% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-alert-zones.R
\encoding{UTF-8}
\name{aemet_alert_zones}
\alias{aemet_alert_zones}
\title{AEMET alert zones}
\source{
\url{https://www.aemet.es/es/eltiempo/prediccion/avisos/ayuda}. See also
Annex 2 and Annex 3 documents, linked from that page.
}
\usage{
aemet_alert_zones(verbose = FALSE, return_sf = FALSE)
}
\arguments{
\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{return_sf}{A logical value. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default), it
returns a \link[tibble:tbl_df]{tibble}. \CRANpkg{sf} must be installed.}
}
\value{
A \link[tibble:tbl_df]{tibble} or an \code{\link[sf:st_sf]{sf}} object.
}
\description{
Retrieves the AEMET geographical zones used for meteorological alerts.
}
\section{Caching}{

The first result retrieved in each session is temporarily cached in
\code{\link[=tempdir]{tempdir()}} to avoid unnecessary requests.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
library(dplyr)
alert_zones <- aemet_alert_zones()
alert_zones

# Cached during this R session.
alert_zones2 <- aemet_alert_zones(verbose = TRUE)

identical(alert_zones, alert_zones2)

# Select and map alert zones.
library(ggplot2)

# Galicia.
alert_zones_sf <- aemet_alert_zones(return_sf = TRUE) |>
  filter(COD_CCAA == "71")

# Coast zones have codes ending in "C".
alert_zones_sf$type <- ifelse(grepl("C$", alert_zones_sf$COD_Z),
  "Coast", "Mainland"
)

ggplot(alert_zones_sf) +
  geom_sf(aes(fill = NOM_PROV)) +
  facet_wrap(~type) +
  scale_fill_brewer(palette = "Blues")
\dontshow{\}) # examplesIf}
}
\seealso{
Weather alerts:
\code{\link[=aemet_alerts]{aemet_alerts()}}

AEMET locations:
\code{\link[=aemet_beaches]{aemet_beaches()}},
\code{\link[=aemet_stations]{aemet_stations()}}
}
\concept{alerts}
\concept{locations}
