% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-alerts.R
\encoding{UTF-8}
\name{aemet_alerts}
\alias{aemet_alerts}
\title{AEMET meteorological alerts}
\source{
\url{https://www.aemet.es/en/eltiempo/prediccion/avisos} and
\url{https://www.aemet.es/es/eltiempo/prediccion/avisos/ayuda} for API status
and alerts reference, including Annex 2 and Annex 3 documentation.
}
\usage{
aemet_alerts(
  ccaa = NULL,
  lang = c("es", "en"),
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{ccaa}{A character vector of autonomous community names or \code{NULL} to
retrieve all autonomous communities.}

\item{lang}{The language of the results, either \code{"es"} (Spanish) or \code{"en"}
(English).}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{return_sf}{A logical value. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default), it
returns a \link[tibble:tbl_df]{tibble}. \CRANpkg{sf} must be installed.}

\item{extract_metadata}{A logical value. If \code{TRUE}, returns a
\link[tibble:tbl_df]{tibble} describing the response fields. See
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{A logical value. If \code{TRUE}, displays a
\code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} unless \code{verbose = TRUE}.}
}
\value{
A \link[tibble:tbl_df]{tibble} or an \code{\link[sf:st_sf]{sf}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Retrieves current meteorological
alerts issued by AEMET.
}
\section{API key}{
Queries to the AEMET OpenData API require an API key. Use \code{\link[=aemet_api_key]{aemet_api_key()}}
to set it globally. Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
# Display CCAA names.
library(dplyr)
aemet_alert_zones() |>
  select(NOM_CCAA) |>
  distinct()

# Base map.
cbasemap <- mapSpain::esp_get_ccaa(ccaa = c(
  "Galicia", "Asturias", "Cantabria",
  "Euskadi"
))

# Alerts.
alerts_north <- aemet_alerts(
  ccaa = c("Galicia", "Asturias", "Cantabria", "Euskadi"),
  return_sf = TRUE
)

# Plot if there are alerts.
if (inherits(alerts_north, "sf")) {
  library(ggplot2)
  library(lubridate)

  alerts_north$day <- date(alerts_north$effective)

  ggplot(alerts_north) +
    geom_sf(data = cbasemap, fill = "grey60") +
    geom_sf(aes(fill = `AEMET-Meteoalerta nivel`)) +
    geom_sf(
      data = cbasemap, fill = "transparent", color = "black",
      linewidth = 0.5
    ) +
    facet_grid(vars(`AEMET-Meteoalerta fenomeno`), vars(day)) +
    scale_fill_manual(values = c(
      "amarillo" = "yellow", naranja = "orange",
      "rojo" = "red"
    ))
}
\dontshow{\}) # examplesIf}
}
\seealso{
See \link[mapSpain:esp_codelist]{mapSpain::esp_codelist} and \code{\link[mapSpain:esp_dict_region_code]{mapSpain::esp_dict_region_code()}} for
autonomous community names.

Weather alerts:
\code{\link[=aemet_alert_zones]{aemet_alert_zones()}}

Weather observations:
\code{\link[=aemet_last_obs]{aemet_last_obs()}}
}
\concept{alerts}
\concept{observations}
