% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-api-key.R
\encoding{UTF-8}
\name{aemet_api_key}
\alias{aemet_api_key}
\title{Install an AEMET OpenData API key}
\usage{
aemet_api_key(apikey, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{apikey}{A character vector of AEMET OpenData API keys. Acquire a key at
\url{https://opendata.aemet.es/centrodedescargas/inicio}. You can install
multiple API keys at once. See \strong{Details}.}

\item{overwrite}{A logical value. If \code{TRUE}, overwrites an existing
\code{AEMET_API_KEY} environment variable.}

\item{install}{A logical value. If \code{TRUE}, installs the key on your local
machine for use in future sessions. Defaults to \code{FALSE}.}
}
\value{
\link[base:NULL]{NULL}, invisibly.
}
\description{
Stores an AEMET OpenData API key on your local machine so it can be used
without including it in your code.
}
\details{
Alternatively, set the key for the current session with
\code{Sys.setenv(AEMET_API_KEY = "Your_Key")}, equivalent to \code{install = FALSE}.
To store it permanently, add \code{AEMET_API_KEY = "Your_Key"} to \code{.Renviron},
equivalent to \code{install = TRUE}.

You can pass multiple \code{apikey} values as a character vector, such as
\code{c(api1, api2)}. In this case, multiple \code{AEMET_API_KEY} values are stored.
In each subsequent API call, \CRANpkg{climaemet} chooses the API key with
the highest remaining quota.

This is useful when performing batch queries to avoid API throttling.
}
\note{
To locate the stored API key, run
\code{tools::R_user_dir("climaemet", "config")}.
}
\section{API key}{
Queries to the AEMET OpenData API require an API key. Use \code{\link[=aemet_api_key]{aemet_api_key()}}
to set it globally. Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
# Do not run these examples.

if (FALSE) {
  aemet_api_key("111111abc", install = TRUE)

  # Check it with:
  Sys.getenv("AEMET_API_KEY")
}

if (FALSE) {
  # Overwrite an existing key:
  aemet_api_key("222222abc", overwrite = TRUE, install = TRUE)

  # Check it with:
  Sys.getenv("AEMET_API_KEY")
}
}
\seealso{
AEMET OpenData API functions:
\code{\link[=aemet_detect_api_key]{aemet_detect_api_key()}},
\code{\link[=get_data_aemet]{get_data_aemet()}}
}
\concept{aemet_api}
\concept{aemet_auth}
