% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-beaches.R
\encoding{UTF-8}
\name{aemet_beaches}
\alias{aemet_beaches}
\title{AEMET beaches}
\usage{
aemet_beaches(verbose = FALSE, return_sf = FALSE)
}
\arguments{
\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{return_sf}{A logical value. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default), it
returns a \link[tibble:tbl_df]{tibble}. \CRANpkg{sf} must be installed.}
}
\value{
A \link[tibble:tbl_df]{tibble} or an \code{\link[sf:st_sf]{sf}} object.
}
\description{
Retrieves the beaches available from the AEMET OpenData API.
}
\section{Caching}{

The first result retrieved in each session is temporarily cached in
\code{\link[=tempdir]{tempdir()}} to avoid unnecessary requests.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
library(dplyr)
beaches <- aemet_beaches()
beaches

# Cached during this R session.
beaches2 <- aemet_beaches(verbose = TRUE)

identical(beaches, beaches2)

# Select and map beaches.
library(ggplot2)
library(mapSpain)

# Alicante / Alacant.
beaches_sf <- aemet_beaches(return_sf = TRUE) |>
  filter(ID_PROVINCIA == "03")

prov <- mapSpain::esp_get_prov("Alicante")

ggplot(prov) +
  geom_sf() +
  geom_sf(
    data = beaches_sf, shape = 4, size = 2.5,
    color = "blue"
  )
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aemet_forecast_beaches]{aemet_forecast_beaches()}}.

AEMET locations:
\code{\link[=aemet_alert_zones]{aemet_alert_zones()}},
\code{\link[=aemet_stations]{aemet_stations()}}
}
\concept{locations}
