% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_valores.R
\name{aemet_daily_clim}
\alias{aemet_daily_clim}
\title{Daily/annual climatology values for a station}
\usage{
aemet_daily_clim(
  station = "all",
  apikey = NULL,
  start = Sys.Date() - lubridate::days(7),
  end = Sys.Date(),
  verbose = FALSE
)
}
\arguments{
\item{station}{Character string with station identifier code(s)
(see \code{\link[=aemet_stations]{aemet_stations()}}) or "all" for all the values.}

\item{apikey}{Character string as personal API key. It can be set
on the environment variable AEMET_API_KEY, see \code{\link[=aemet_api_key]{aemet_api_key()}}.}

\item{start}{Character string as start date (format: \%Y\%m\%d).}

\item{end}{Character string as end date (format: \%Y\%m\%d).}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}
}
\value{
a tibble
}
\description{
Get daily climatology values for a station for a maximum period of one year.
}
\note{
The API has some limits on large periods (31 days for "all" and
one year for individual stations). For larger period consider use
\code{\link[=aemet_daily_period]{aemet_daily_period()}}/\code{\link[=aemet_daily_period_all]{aemet_daily_period_all()}}, that are optimized for
batch downloading.
}
\examples{

# Run this example only if AEMET_API_KEY is set
apikey <- Sys.getenv("AEMET_API_KEY")
if (apikey != "") {
  library(tibble)
  obs <- aemet_daily_clim("9434", apikey, "2021-01-07", "2021-01-10")
  glimpse(obs)
}
}
\seealso{
\code{\link[=aemet_daily_period]{aemet_daily_period()}}, \code{\link[=aemet_daily_period_all]{aemet_daily_period_all()}}
}
\concept{aemet_valores}
