% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-api-key.R
\encoding{UTF-8}
\name{aemet_detect_api_key}
\alias{aemet_detect_api_key}
\alias{aemet_show_api_key}
\title{Check for an AEMET OpenData API key}
\usage{
aemet_detect_api_key(...)

aemet_show_api_key(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
\code{\link[=aemet_detect_api_key]{aemet_detect_api_key()}} returns a \link[base:logical]{logical}
value, \code{TRUE} if an API key is available and \code{FALSE} otherwise.
\code{\link[=aemet_show_api_key]{aemet_show_api_key()}} returns a \link[base:character]{character} vector
containing the available API keys.
}
\description{
Detects whether an API key is available in the current session. An existing
environment variable is preserved. Otherwise, a key stored permanently with
\code{\link[=aemet_api_key]{aemet_api_key()}} is loaded.
}
\examples{

aemet_detect_api_key()

# Caution: This may reveal API keys.
if (FALSE) {
  aemet_show_api_key()
}
}
\seealso{
AEMET OpenData API functions:
\code{\link[=aemet_api_key]{aemet_api_key()}},
\code{\link[=get_data_aemet]{get_data_aemet()}}
}
\concept{aemet_api}
\concept{aemet_auth}
