% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-monthly.R
\encoding{UTF-8}
\name{aemet_monthly_clim}
\alias{aemet_monthly_clim}
\alias{aemet_monthly_period}
\alias{aemet_monthly_period_all}
\title{Monthly and annual climatology values}
\usage{
aemet_monthly_clim(
  station = NULL,
  year = as.integer(format(Sys.Date(), "\%Y")),
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_monthly_period(
  station = NULL,
  start = as.integer(format(Sys.Date(), "\%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_monthly_period_all(
  start = as.integer(format(Sys.Date(), "\%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{station}{A character vector of station identifiers. See
\code{\link[=aemet_stations]{aemet_stations()}}.}

\item{year}{A numeric year in \code{YYYY} format.}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{return_sf}{A logical value. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default), it
returns a \link[tibble:tbl_df]{tibble}. \CRANpkg{sf} must be installed.}

\item{extract_metadata}{A logical value. If \code{TRUE}, returns a
\link[tibble:tbl_df]{tibble} describing the response fields. See
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{A logical value. If \code{TRUE}, displays a
\code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} unless \code{verbose = TRUE}.}

\item{start}{A numeric value specifying the start year in \code{YYYY} format.}

\item{end}{A numeric value specifying the end year in \code{YYYY} format.}
}
\value{
A \link[tibble:tbl_df]{tibble} or an \code{\link[sf:st_sf]{sf}} object.
}
\description{
Retrieves monthly or annual climatology values for one or more stations.
\code{\link[=aemet_monthly_period]{aemet_monthly_period()}} and \code{\link[=aemet_monthly_period_all]{aemet_monthly_period_all()}} allow requests
that span several years.
}
\section{API key}{
Queries to the AEMET OpenData API require an API key. Use \code{\link[=aemet_api_key]{aemet_api_key()}}
to set it globally. Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
obs <- aemet_monthly_clim(station = c("9434", "3195"), year = 2000)
dplyr::glimpse(obs)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aemet_stations]{aemet_stations()}} for station identifiers.

Climatology:
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link[=aemet_normal_clim]{aemet_normal_clim()}}
}
\concept{climatology}
