% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{aemet_monthly_clim}
\alias{aemet_monthly_clim}
\title{Monthly/annual climatology values for a station}
\usage{
aemet_monthly_clim(station, apikey, year)
}
\arguments{
\item{station}{Station identifier code (see \code{\link{aemet_stations}}).}

\item{apikey}{Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).}

\item{year}{a date (format: \%Y)}
}
\value{
a data.frame
}
\description{
Get monthly/annual climatology values for a station
}
\examples{
\dontrun{
aemet_monthly_clim("9434", apikey, 2000)
}
}
