% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{aemet_monthly_period}
\alias{aemet_monthly_period}
\title{Monthly climatology values for a station}
\usage{
aemet_monthly_period(station, apikey, start, end)
}
\arguments{
\item{station}{Station identifier code (see \code{\link{aemet_stations}}).}

\item{apikey}{Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).}

\item{start}{a start year (format: \%Y)}

\item{end}{a end year (format: \%Y)}
}
\value{
a data.frame
}
\description{
Get monthly climatology values for a period of years for a station
}
\examples{
\dontrun{
aemet_monthly_period("9434", apikey, 2000, 2010)
}

}
