% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{aemet_normal_clim}
\alias{aemet_normal_clim}
\title{Normal climatology values for a station}
\usage{
aemet_normal_clim(station, apikey)
}
\arguments{
\item{station}{Character string with station identifier code(s)
(see \code{\link[=aemet_stations]{aemet_stations()}}) or "all" for all the values.}

\item{apikey}{Character string as personal API key. It can be set
on the environment variable AEMET_API_KEY, see \code{\link[=aemet_api_key]{aemet_api_key()}}.}
}
\value{
a tibble
}
\description{
Get normal climatology values for a station. Standard climatology from 1981 to 2010.
}
\note{
Code modified from project https://github.com/SevillaR/aemet
}
\examples{
\dontrun{
aemet_normal_clim("9434", apikey)
}

}
