% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{aemet_normal_clim_all}
\alias{aemet_normal_clim_all}
\title{Normal climatology values for all stations}
\usage{
aemet_normal_clim_all(apikey)
}
\arguments{
\item{apikey}{Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).}
}
\value{
a data.frame
}
\description{
Get normal climatology values for all stations
}
\examples{
\dontrun{
aemet_normal_clim_all(apikey)
}

}
