% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climaemet-package.R
\docType{package}
\encoding{UTF-8}
\name{climaemet-package}
\alias{climaemet}
\alias{climaemet-package}
\title{climaemet: Access and Visualize AEMET Weather and Climate Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides access to meteorological observations, forecasts, alerts and climatology data from the Spanish Meteorological Agency (AEMET) through the 'AEMET OpenData' API \url{https://opendata.aemet.es/}. Includes tools for working with tabular and spatial data and for creating Walter-Lieth climate diagrams, warming stripes and wind roses.
}
\references{
Walter H, Lieth H (1967). \emph{Klimadiagramm-Weltatlas}. VEB Gustav Fischer
Verlag, Jena. Published in three installments, 1960-1967,
\url{https://hdl.handle.net/2268.1/7079}.

Guijarro JA (2026). \emph{climatol: Climate Tools (Series Homogenization and
Derived Products)}. \doi{10.32614/CRAN.package.climatol}. R package version
4.5-0, \url{https://CRAN.R-project.org/package=climatol}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/climaemet/}
  \item \url{https://github.com/rOpenSpain/climaemet}
  \item Report bugs at \url{https://github.com/rOpenSpain/climaemet/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})
  \item Manuel Pizarro (\href{https://orcid.org/0000-0002-6981-0154}{ORCID}) [copyright holder]
  \item Gema Fernández-Avilés (\href{https://orcid.org/0000-0001-5934-1916}{ORCID})
}

Other contributors:
\itemize{
  \item AEMET (\href{https://ror.org/04kxf1r09}{ROR}) [copyright holder]
}

}
\keyword{internal}
