% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climaemet-news.R
\encoding{UTF-8}
\name{climaemet_news}
\alias{climaemet_news}
\title{Show the latest \CRANpkg{climaemet} news}
\usage{
climaemet_news()
}
\value{
\link[base:NULL]{NULL}, invisibly. This function is called for its
side effect.
}
\description{
Opens the \code{NEWS} file for \CRANpkg{climaemet}.
}
\examples{
\dontrun{
climaemet_news()
}

}
\seealso{
Helper functions:
\code{\link[=dms2decdegrees]{dms2decdegrees()}},
\code{\link[=first_day_of_year]{first_day_of_year()}}
}
\concept{helpers}
