% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatogram.R
\encoding{UTF-8}
\name{climatogram_normal}
\alias{climatogram_normal}
\title{Walter-Lieth climate diagram from climatological normals}
\usage{
climatogram_normal(
  station,
  labels = "en",
  verbose = FALSE,
  ggplot2 = TRUE,
  ...
)
}
\arguments{
\item{station}{A character string containing a station identifier. See
\code{\link[=aemet_stations]{aemet_stations()}}.}

\item{labels}{A character string specifying the language for the x-axis
month labels, such as \code{"en"} (English), \code{"es"} (Spanish) or \code{"fr"}
(French).}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{ggplot2}{A logical value. If \code{TRUE}, the function uses
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}. If \code{FALSE}, it uses \code{\link[climatol:diagwl]{climatol::diagwl()}}.}

\item{...}{Further arguments passed to
\code{\link[climatol:diagwl]{climatol::diagwl()}} or \code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}, depending on the
value of \code{ggplot2}.}
}
\value{
A plot produced by \code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}} or
\code{\link[climatol:diagwl]{climatol::diagwl()}}, depending on \code{ggplot2}.
}
\description{
Plots a Walter-Lieth climate diagram from climatological normal values for a
station. The diagram summarizes local climate conditions for 1981–2010.
}
\section{API key}{
Queries to the AEMET OpenData API require an API key. Use \code{\link[=aemet_api_key]{aemet_api_key()}}
to set it globally. Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
climatogram_normal("9434")
\dontshow{\}) # examplesIf}
}
\references{
Walter H, Lieth H (1967). \emph{Klimadiagramm-Weltatlas}. VEB Gustav Fischer
Verlag, Jena. Published in three installments, 1960-1967,
\url{https://hdl.handle.net/2268.1/7079}.

Guijarro JA (2026). \emph{climatol: Climate Tools (Series Homogenization and
Derived Products)}. \doi{10.32614/CRAN.package.climatol}. R package version
4.5-0, \url{https://CRAN.R-project.org/package=climatol}.
}
\seealso{
\link{climaemet_9434_climatogram}.

Walter-Lieth climate diagrams:
\code{\link[=climatogram_period]{climatogram_period()}},
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}
}
\concept{climatogram}
