% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\encoding{UTF-8}
\name{first_day_of_year}
\alias{first_day_of_year}
\alias{last_day_of_year}
\title{First and last day of a year}
\usage{
first_day_of_year(year = NULL)

last_day_of_year(year = NULL)
}
\arguments{
\item{year}{A numeric year in \code{YYYY} format.}
}
\value{
A \link[base:character]{character} string containing a date in
\code{YYYY-MM-DD} format.
}
\description{
Returns the first or last calendar day of a year.
}
\examples{
first_day_of_year(2000)
last_day_of_year(2020)
}
\seealso{
Helper functions:
\code{\link[=climaemet_news]{climaemet_news()}},
\code{\link[=dms2decdegrees]{dms2decdegrees()}}
}
\concept{helpers}
