% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{first_day_of_year}
\alias{first_day_of_year}
\title{First day of year}
\usage{
first_day_of_year(year, month = 1)
}
\arguments{
\item{year}{a start year (format: \%Y)}

\item{month}{a start year (format: \%m)}
}
\value{
date (format: \%Y\%m\%d)
}
\description{
Get first day of year
}
\examples{
\donttest{
first_day_of_year(2000)
}

}
