% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-api-query.R
\encoding{UTF-8}
\name{get_data_aemet}
\alias{get_data_aemet}
\alias{get_metadata_aemet}
\title{Query the AEMET OpenData API}
\source{
\url{https://opendata.aemet.es/dist/index.html}.
}
\usage{
get_data_aemet(apidest, verbose = FALSE)

get_metadata_aemet(apidest, verbose = FALSE)
}
\arguments{
\item{apidest}{A character string containing the destination URL. See
\url{https://opendata.aemet.es/dist/index.html}.}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}
}
\value{
A \link[tibble:tbl_df]{tibble} when possible, otherwise a \link[base:raw]{raw}
vector or a \link[base:character]{character} string as provided by
\code{\link[httr2:resp_body_raw]{httr2::resp_body_raw()}} or \code{\link[httr2:resp_body_string]{httr2::resp_body_string()}}.
}
\description{
Retrieves data and metadata from AEMET and converts JSON responses to a
\link[tibble:tbl_df]{tibble} when possible.
}
\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
# Run only when AEMET_API_KEY is detected.

url <- "/api/valores/climatologicos/inventarioestaciones/todasestaciones"

get_data_aemet(url)

# Metadata.

get_metadata_aemet(url)

# Get data from any API endpoint.

# Plain text.

plain <- get_data_aemet("/api/prediccion/nacional/hoy")

cat(plain)

# An image.

image <- get_data_aemet("/api/mapasygraficos/analisis")

# Write and read.
tmp <- tempfile(fileext = ".gif")

writeBin(image, tmp)

gganimate::gif_file(tmp)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{vignette("extending-climaemet", package = "climaemet")} provides usage
examples.

AEMET OpenData API functions:
\code{\link[=aemet_api_key]{aemet_api_key()}},
\code{\link[=aemet_detect_api_key]{aemet_detect_api_key()}}
}
\concept{aemet_api}
\concept{aemet_low}
