% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\encoding{UTF-8}
\name{ggwindrose}
\alias{ggwindrose}
\title{Plot a wind rose}
\usage{
ggwindrose(
  speed,
  direction,
  n_directions = 8,
  n_speeds = 5,
  speed_cuts = NA,
  col_pal = "GnBu",
  legend_title = "Wind speed (m/s)",
  calm_wind = 0,
  n_col = 1,
  facet = NULL,
  plot_title = "",
  stack_reverse = FALSE,
  ...
)
}
\arguments{
\item{speed}{A numeric vector of wind speeds.}

\item{direction}{A numeric vector of wind directions.}

\item{n_directions}{The number of direction bins (petals) to plot. Valid
values are \code{4}, \code{8} or \code{16}.}

\item{n_speeds}{The number of equally spaced wind speed bins to plot when
\code{speed_cuts} is \code{NA}. Defaults to \code{5}.}

\item{speed_cuts}{A numeric vector with the cut points for the wind speed
intervals or \code{NA} (default).}

\item{col_pal}{A character string specifying an \code{\link[=hcl.pals]{hcl.pals()}} color palette.}

\item{legend_title}{A character string or expression for the legend title.}

\item{calm_wind}{The upper wind speed limit considered calm. Defaults to \code{0}.}

\item{n_col}{The number of plot columns. Defaults to \code{1}.}

\item{facet}{A character or factor vector of facets used to plot wind roses.}

\item{plot_title}{A character string for the plot title.}

\item{stack_reverse}{A logical value. If \code{TRUE}, reverses the stack order of
speed cuts. See \strong{Examples}.}

\item{...}{Further arguments (ignored).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plots a wind rose showing wind speed and direction with \CRANpkg{ggplot2}.
}
\examples{
library(ggplot2)

speed <- climaemet::climaemet_9434_wind$velmedia
direction <- climaemet::climaemet_9434_wind$dir

rose <- ggwindrose(
  speed = speed,
  direction = direction,
  speed_cuts = seq(0, 16, 4),
  legend_title = "Wind speed (m/s)",
  calm_wind = 0,
  n_col = 1,
  plot_title = "Zaragoza Airport"
)
rose + labs(
  subtitle = "2000-2020",
  caption = "Source: AEMET"
)

# Reverse the stack.

ggwindrose(
  speed = speed,
  direction = direction,
  speed_cuts = seq(0, 16, 4),
  legend_title = "Wind speed (m/s)",
  calm_wind = 0,
  n_col = 1,
  plot_title = "Zaragoza Airport",
  stack_reverse = TRUE
) +
  labs(
    subtitle = "2000-2020",
    caption = "Source: AEMET"
  )

}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}} for additional arguments to pass to
\code{\link[=ggwindrose]{ggwindrose()}} and \link{climaemet_9434_wind}.

Wind roses:
\code{\link[=windrose_days]{windrose_days()}},
\code{\link[=windrose_period]{windrose_period()}}
}
\concept{wind}
