% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{last_day_of_year}
\alias{last_day_of_year}
\title{Last day of year}
\usage{
last_day_of_year(year)
}
\arguments{
\item{year}{Numeric value as year (format: \%Y).}
}
\value{
Character string as date (format: \%Y\%m\%d).
}
\description{
Get last day of year.
}
\examples{
\donttest{
last_day_of_year(2000)
}

}
