% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\encoding{UTF-8}
\name{windrose_period}
\alias{windrose_period}
\title{Wind rose for a range of years}
\usage{
windrose_period(
  station,
  start = 2000,
  end = 2010,
  n_directions = 8,
  n_speeds = 5,
  speed_cuts = NA,
  col_pal = "GnBu",
  calm_wind = 0,
  legend_title = "Wind speed (m/s)",
  verbose = FALSE
)
}
\arguments{
\item{station}{A character string containing a station identifier. See
\code{\link[=aemet_stations]{aemet_stations()}}.}

\item{start}{A numeric value specifying the start year in \code{YYYY} format.}

\item{end}{A numeric value specifying the end year in \code{YYYY} format.}

\item{n_directions}{The number of direction bins (petals) to plot. Valid
values are \code{4}, \code{8} or \code{16}.}

\item{n_speeds}{The number of equally spaced wind speed bins to plot when
\code{speed_cuts} is \code{NA}. Defaults to \code{5}.}

\item{speed_cuts}{A numeric vector with the cut points for the wind speed
intervals or \code{NA} (default).}

\item{col_pal}{A character string specifying an \code{\link[=hcl.pals]{hcl.pals()}} color palette.}

\item{calm_wind}{The upper wind speed limit considered calm. Defaults to \code{0}.}

\item{legend_title}{A character string or expression for the legend title.}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plots a wind rose showing wind speed and direction at a station over a
time period.
}
\section{API key}{
Queries to the AEMET OpenData API require an API key. Use \code{\link[=aemet_api_key]{aemet_api_key()}}
to set it globally. Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
\donttest{
# Do not run this example.
if (FALSE) {
  # Downloading data may take a few minutes.
  windrose_period("9434",
    start = 2000, end = 2010,
    speed_cuts = 4
  )
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \code{\link[=aemet_daily_period]{aemet_daily_period()}} retrieves daily climatology data by period.
\item \link{climaemet_9434_wind} provides example wind observations.
}

Wind roses:
\code{\link[=ggwindrose]{ggwindrose()}},
\code{\link[=windrose_days]{windrose_days()}}
}
\concept{wind}
