% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-alert-zones.R
\encoding{UTF-8}
\name{aemet_alert_zones}
\alias{aemet_alert_zones}
\title{AEMET alert zones}
\source{
\url{https://www.aemet.es/es/eltiempo/prediccion/avisos/ayuda}. See also
Annex 2 and Annex 3 documents, linked from that page.
}
\usage{
aemet_alert_zones(verbose = FALSE, return_sf = FALSE)
}
\arguments{
\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default value), it
returns a \link[tibble:tbl_df]{tibble}. The \CRANpkg{sf} package must be
installed.}
}
\value{
A \link[tibble:tbl_df]{tibble} or a \CRANpkg{sf} object.
}
\description{
Get AEMET alert zones.
}
\details{
The first result of each call per session is temporarily cached in
\code{\link[=tempdir]{tempdir()}} to avoid unnecessary API calls.
}
\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
library(tibble)
alert_zones <- aemet_alert_zones()
alert_zones

# Cached during this R session
alert_zones2 <- aemet_alert_zones(verbose = TRUE)

identical(alert_zones, alert_zones2)

# Select and map alert zones.
library(dplyr)
library(ggplot2)

# Galicia
alert_zones_sf <- aemet_alert_zones(return_sf = TRUE) |>
  filter(COD_CCAA == "71")

# Coast zones are identified by a "C" in COD_Z.
alert_zones_sf$type <- ifelse(grepl("C$", alert_zones_sf$COD_Z),
  "Coast", "Mainland"
)

ggplot(alert_zones_sf) +
  geom_sf(aes(fill = NOM_PROV)) +
  facet_wrap(~type) +
  scale_fill_brewer(palette = "Blues")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aemet_alerts]{aemet_alerts()}}

AEMET data functions:
\code{\link[=aemet_alerts]{aemet_alerts()}},
\code{\link[=aemet_beaches]{aemet_beaches()}},
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link[=aemet_forecast_beaches]{aemet_forecast_beaches()}},
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_fires]{aemet_forecast_fires()}},
\code{\link[=aemet_last_obs]{aemet_last_obs()}},
\code{\link{aemet_monthly}},
\code{\link{aemet_normal}},
\code{\link[=aemet_stations]{aemet_stations()}}
}
\concept{aemet_api_data}
