% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-alert-zones.R
\encoding{UTF-8}
\name{aemet_alert_zones}
\alias{aemet_alert_zones}
\title{AEMET alert zones}
\source{
\url{https://www.aemet.es/es/eltiempo/prediccion/avisos/ayuda}. See also
Annex 2 and Annex 3 documents, linked from that page.
}
\usage{
aemet_alert_zones(verbose = FALSE, return_sf = FALSE)
}
\arguments{
\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default value), it
returns a \link[dplyr:tibble]{tibble}. The \CRANpkg{sf} package must be
installed.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object.
}
\description{
Retrieves the AEMET geographical zones used for meteorological alerts.
}
\details{
The first result of each call per session is temporarily cached in
\code{\link[=tempdir]{tempdir()}} to avoid unnecessary API calls.
}
\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
library(dplyr)
alert_zones <- aemet_alert_zones()
alert_zones

# Cached during this R session.
alert_zones2 <- aemet_alert_zones(verbose = TRUE)

identical(alert_zones, alert_zones2)

# Select and map alert zones.
library(ggplot2)

# Galicia.
alert_zones_sf <- aemet_alert_zones(return_sf = TRUE) |>
  filter(COD_CCAA == "71")

# Coast zones have codes ending in "C".
alert_zones_sf$type <- ifelse(grepl("C$", alert_zones_sf$COD_Z),
  "Coast", "Mainland"
)

ggplot(alert_zones_sf) +
  geom_sf(aes(fill = NOM_PROV)) +
  facet_wrap(~type) +
  scale_fill_brewer(palette = "Blues")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aemet_alerts]{aemet_alerts()}}

Observations and locations:
\code{\link[=aemet_alerts]{aemet_alerts()}},
\code{\link[=aemet_beaches]{aemet_beaches()}},
\code{\link[=aemet_last_obs]{aemet_last_obs()}},
\code{\link[=aemet_stations]{aemet_stations()}}
}
\concept{observations}
