% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_alerts.R
\name{aemet_alerts}
\alias{aemet_alerts}
\title{AEMET Meteorological warnings}
\source{
\url{https://www.aemet.es/en/eltiempo/prediccion/avisos}.
}
\usage{
aemet_alerts(ccaa = NULL, lang = c("es", "en"), verbose = FALSE)
}
\arguments{
\item{ccaa}{A vector of names for autonomous communities or \code{NULL} to get all
the autonomous communities.}

\item{lang}{Language of the results. It can be \code{"es"} (Spanish) or \code{"en"}
(English).}

\item{verbose}{Logical \code{TRUE/FALSE}. Provides information about the flow of
information between the client and server.}
}
\value{
A \CRANpkg{sf} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Get a database of current meteorological
alerts.
}
\examples{
\dontshow{if (aemet_detect_api_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Display names of CCAAs
unique(mapSpain::esp_codelist$ine.ccaa.name)

# Base map
cbasemap <- mapSpain::esp_get_ccaa(ccaa = c(
  "Galicia",
  "Asturias, Principado de"
))

# Alerts
alerts_north <- aemet_alerts(ccaa = c("Galicia", "Asturias, Principado de"))

# If any alert
if (inherits(alerts_north, "sf")) {
  library(ggplot2)
  library(lubridate)


  alerts_north$day <- date(alerts_north$effective)

  ggplot(alerts_north) +
    geom_sf(data = cbasemap, fill = "grey60") +
    geom_sf(aes(fill = `AEMET-Meteoalerta nivel`)) +
    geom_sf(
      data = cbasemap, fill = "transparent", color = "black",
      linewidth = 0.5
    ) +
    facet_wrap(vars(day, `AEMET-Meteoalerta fenomeno`)) +
    scale_fill_manual(values = c(
      "amarillo" = "yellow", naranja = "orange",
      "rojo" = "red"
    ))
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[mapSpain:esp_get_ccaa]{mapSpain::esp_get_ccaa()}}. See also \link[mapSpain:esp_codelist]{mapSpain::esp_codelist},
\code{\link[mapSpain:esp_dict]{mapSpain::esp_dict_region_code()}} and \strong{Examples} to get the names of the
autonomous communities.

Other aemet_api_data: 
\code{\link{aemet_beaches}()},
\code{\link{aemet_daily_clim}()},
\code{\link{aemet_extremes_clim}()},
\code{\link{aemet_forecast_beaches}()},
\code{\link{aemet_forecast_daily}()},
\code{\link{aemet_last_obs}()},
\code{\link{aemet_monthly}},
\code{\link{aemet_normal}},
\code{\link{aemet_stations}()}
}
\concept{aemet_api_data}
