% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-api-key.R
\name{aemet_api_key}
\alias{aemet_api_key}
\title{Install an AEMET API Key}
\usage{
aemet_api_key(apikey, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{apikey}{The API key provided to you from the AEMET formatted in quotes.
A key can be acquired at
\url{https://opendata.aemet.es/centrodedescargas/inicio}. You can install
several API keys as a character vector; see \strong{Details}.}

\item{overwrite}{If \code{TRUE}, overwrites an existing
\code{AEMET_API_KEY} already set on your local machine.}

\item{install}{If \code{TRUE}, installs the key on your local machine for
use in future sessions. Defaults to \code{FALSE}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
This function will store your AEMET API key on your local machine so it can
be called securely without being stored in your code.

Alternatively, you can install the API key manually:
\itemize{
\item Run \code{Sys.setenv(AEMET_API_KEY = "Your_Key")}. You will need to run this
command in each session (similar to \code{install = FALSE}).
\item Write this line in your .Renviron file: \code{AEMET_API_KEY = "Your_Key"}
(same behavior as \code{install = TRUE}). This stores your API key
permanently.
}
}
\details{
You can pass several \code{apikey} values as a character vector \code{c(api1, api2)};
in this case, multiple \code{AEMET_API_KEY} values are generated. In each
subsequent API call, \CRANpkg{climaemet} chooses the API key with the highest
remaining quota.

This is useful when performing batch queries to avoid API throttling.
}
\note{
To locate your API Key on your local machine, run
\code{rappdirs::user_cache_dir("climaemet", "R")}.
}
\examples{
# Don't run these examples!

if (FALSE) {
  aemet_api_key("111111abc", install = TRUE)

  # You can check it with:
  Sys.getenv("AEMET_API_KEY")
}

if (FALSE) {
  # If you need to overwrite an existing key:
  aemet_api_key("222222abc", overwrite = TRUE, install = TRUE)

  # You can check it with:
  Sys.getenv("AEMET_API_KEY")
}
}
\seealso{
Other aemet_auth: 
\code{\link{aemet_detect_api_key}()}
}
\concept{aemet_auth}
