% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aemet_api_key}
\alias{aemet_api_key}
\title{Install a AEMET API Key in Your \code{.Renviron} File for Repeated Use}
\usage{
aemet_api_key(apikey, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{apikey}{The API key provided to you from the AEMET formatted in quotes.
A key can be acquired at
\url{https://opendata.aemet.es/centrodedescargas/inicio}.}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing
AEMET_API_KEY that you already have in your \code{.Renviron} file.}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for
use in future sessions.  Defaults to FALSE.}
}
\description{
This function will add your AEMET API key to your \code{.Renviron} file so it can
be called securely without being stored in your code. After you have
installed your key, it can be called any time by typing
\code{Sys.getenv("AEMET_API_KEY")} and can be
used in package functions by simply typing AEMET_API_KEY If you do not have
an \code{.Renviron} file, the function will create on for you.
If you already have an \code{.Renviron} file, the function will append the key to
your existing file, while making a backup of your
original file for disaster recovery purposes.
}
\note{
Code adapted from
\url{https://walker-data.com/tidycensus/reference/census_api_key.html}
}
\examples{

# Don't run these examples!

if (FALSE) {
  aemet_api_key("111111abc", install = TRUE)
  # First time, reload your environment so you can use the key without restarting R.
  readRenviron("~/.Renviron")
  # You can check it with:
  Sys.getenv("AEMET_API_KEY")
}

if (FALSE) {
  # If you need to overwrite an existing key:
  aemet_api_key("111111abc", overwrite = TRUE, install = TRUE)
  readRenviron("~/.Renviron")
  # You can check it with:
  Sys.getenv("AEMET_API_KEY")
}
}
\concept{helpers}
