% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-beaches.R
\encoding{UTF-8}
\name{aemet_beaches}
\alias{aemet_beaches}
\title{AEMET beaches}
\usage{
aemet_beaches(verbose = FALSE, return_sf = FALSE)
}
\arguments{
\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default value), it
returns a \link[dplyr:tibble]{tibble}. The \CRANpkg{sf} package must be
installed.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object.
}
\description{
Retrieves the beaches available from the AEMET API.
}
\details{
The first result of the API call in each session is temporarily cached in
\code{\link[=tempdir]{tempdir()}} to avoid unnecessary API calls.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
library(dplyr)
beaches <- aemet_beaches()
beaches

# Cached during this R session.
beaches2 <- aemet_beaches(verbose = TRUE)

identical(beaches, beaches2)

# Select and map beaches.
library(ggplot2)
library(mapSpain)

# Alicante / Alacant.
beaches_sf <- aemet_beaches(return_sf = TRUE) |>
  filter(ID_PROVINCIA == "03")

prov <- mapSpain::esp_get_prov("Alicante")

ggplot(prov) +
  geom_sf() +
  geom_sf(
    data = beaches_sf, shape = 4, size = 2.5,
    color = "blue"
  )
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aemet_forecast_beaches]{aemet_forecast_beaches()}}

AEMET data functions:
\code{\link[=aemet_alert_zones]{aemet_alert_zones()}},
\code{\link[=aemet_alerts]{aemet_alerts()}},
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link[=aemet_forecast_beaches]{aemet_forecast_beaches()}},
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_fires]{aemet_forecast_fires()}},
\code{\link[=aemet_last_obs]{aemet_last_obs()}},
\code{\link{aemet_monthly}},
\code{\link{aemet_normal}},
\code{\link[=aemet_stations]{aemet_stations()}}
}
\concept{aemet_api_data}
