% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-daily.R
\encoding{UTF-8}
\name{aemet_daily_clim}
\alias{aemet_daily_clim}
\alias{aemet_daily}
\alias{aemet_daily_period}
\alias{aemet_daily_period_all}
\title{Daily and annual climatology values}
\usage{
aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_daily_period(
  station,
  start = as.integer(format(Sys.Date(), "\%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_daily_period_all(
  start = as.integer(format(Sys.Date(), "\%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{station}{Character string with station identifier code(s) (see
\code{\link[=aemet_stations]{aemet_stations()}}) or \code{"all"} for all the stations.}

\item{start, end}{Character strings with start and end dates. See
\strong{Details}.}

\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default value), it
returns a \link[dplyr:tibble]{tibble}. The \CRANpkg{sf} package must be
installed.}

\item{extract_metadata}{Logical. If \code{TRUE}, the output is a
\link[dplyr:tibble]{tibble} with the description of the fields. See also
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{Logical. Displays a \code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} object. If
\code{verbose = TRUE}, it will not be displayed.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object.
}
\description{
Retrieves climatology values for one station or all available stations.
\code{aemet_daily_period()} and \code{aemet_daily_period_all()} are shortcuts for
\code{aemet_daily_clim()}.
}
\details{
For \code{aemet_daily_clim()}, \code{start} and \code{end} must be \code{Date} objects or
strings in \code{YYYY-MM-DD} format, such as \code{"2020-12-31"}, that can be coerced
with \code{\link[=as.Date]{as.Date()}}. For \code{aemet_daily_period()} and
\code{aemet_daily_period_all()}, they must be strings representing the years to
extract, such as \code{"2018"} and \code{"2020"}.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}

library(dplyr)
obs <- aemet_daily_clim(c("9434", "3195"))
glimpse(obs)

# Metadata.
meta <- aemet_daily_clim(c("9434", "3195"), extract_metadata = TRUE)

glimpse(meta$campos)
\dontshow{\}) # examplesIf}
}
\seealso{
Climatology:
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link{aemet_monthly}},
\code{\link{aemet_normal}}
}
\concept{climatology}
