% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_daily.R
\name{aemet_daily_clim}
\alias{aemet_daily_clim}
\alias{aemet_daily}
\alias{aemet_daily_period}
\alias{aemet_daily_period_all}
\title{Daily/annual climatology values}
\usage{
aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period(
  station,
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period_all(
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)
}
\arguments{
\item{station}{Character string with station identifier code(s)
(see \code{\link[=aemet_stations]{aemet_stations()}}) or "all" for all the stations.}

\item{start, end}{Character string with start and end date. See Details.}

\item{verbose}{Logical \code{TRUE/FALSE}. Provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical \code{TRUE} or \code{FALSE}.
Should the function return an \code{sf} spatial object? If \code{FALSE}
(the default value) it returns a tibble. Note that you need to
have the \code{sf} package installed.}
}
\value{
A tibble or a \code{sf} object
}
\description{
Get climatology values for a station or for all the available stations.
Note that \code{aemet_daily_period()} and \code{aemet_daily_period_all()} are shortcuts
of \code{aemet_daily_clim()}.
}
\details{
\code{start} and \code{end} parameters should be:
\itemize{
\item For \code{aemet_daily_clim()}: A \code{Date} object or a string with format:
YYYY-MM-DD (2020-12-31).
\item For \code{aemet_daily_period()} and \code{aemet_daily_period_all()}: A string
representing the year(s) to be extracted: "2020", "2018".
}
}
\section{API Key}{
You need to set your API Key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
}

\examples{
# Run this example only if AEMET_API_KEY is detected

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_daily_clim(c("9434", "3195"))
  glimpse(obs)
}
}
\seealso{
\code{\link[=aemet_api_key]{aemet_api_key()}}
}
\concept{aemet_api_data}
