% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_valores.R
\name{aemet_daily_period}
\alias{aemet_daily_period}
\title{Daily climatology values of a station for a time period}
\usage{
aemet_daily_period(
  station,
  apikey = NULL,
  start = 2020,
  end = 2020,
  verbose = FALSE
)
}
\arguments{
\item{station}{Character string as station identifier code
(see \code{\link[=aemet_stations]{aemet_stations()}}).}

\item{apikey}{Character string as personal API key. It can be set
on the environment variable AEMET_API_KEY, see \code{\link[=aemet_api_key]{aemet_api_key()}}.}

\item{start}{Numeric value as start year (format: \%Y).}

\item{end}{Numeric value as end year (format: \%Y).}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}
}
\value{
a tibble
}
\description{
Get daily climatology values for a period of years for a station.
}
\examples{
\dontrun{
aemet_daily_period("9434", apikey, 2000, 2010)
}

}
\seealso{
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}}
}
\concept{aemet_valores}
