% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{aemet_daily_period_all}
\alias{aemet_daily_period_all}
\title{Daily climatology values for all stations}
\usage{
aemet_daily_period_all(apikey, start, end)
}
\arguments{
\item{apikey}{Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).}

\item{start}{a start year (format: \%Y)}

\item{end}{a end year (format: \%Y)}
}
\value{
a data.frame
}
\description{
Get daily climatology values for a period of years for all stations
}
\examples{
\dontrun{
aemet_daily_period_all(apikey, 2000, 2010)
}

}
