% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-api-key.R
\encoding{UTF-8}
\name{aemet_detect_api_key}
\alias{aemet_detect_api_key}
\alias{aemet_show_api_key}
\title{Check for an AEMET API key}
\usage{
aemet_detect_api_key(...)

aemet_show_api_key(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
\code{TRUE} if an API key is available and \code{FALSE} otherwise.
\code{aemet_show_api_key()} displays stored API keys.
}
\description{
Detects whether an API key is available in the current session. An existing
environment variable is preserved. Otherwise, a key stored permanently with
\code{\link[=aemet_api_key]{aemet_api_key()}} is loaded.
}
\examples{

aemet_detect_api_key()

# CAUTION: This may reveal API keys.
if (FALSE) {
  aemet_show_api_key()
}
}
\seealso{
AEMET API authentication:
\code{\link[=aemet_api_key]{aemet_api_key()}}
}
\concept{aemet_auth}
