% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-api-key.R
\name{aemet_detect_api_key}
\alias{aemet_detect_api_key}
\alias{aemet_show_api_key}
\title{Check if an AEMET API Key is present for the current session}
\usage{
aemet_detect_api_key(...)

aemet_show_api_key(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
\code{TRUE} or \code{FALSE}. \code{aemet_show_api_key()} displays your stored API keys.
}
\description{
Detects whether an API key is available in the current session:
\itemize{
\item If an API key is already set as an environment variable, it is
preserved.
\item If no environment variable is set and an API key has been stored
permanently via \code{\link[=aemet_api_key]{aemet_api_key()}}, it is loaded.
}
}
\examples{

aemet_detect_api_key()

# CAUTION: This may reveal API Keys
if (FALSE) {
  aemet_show_api_key()
}
}
\seealso{
Other aemet_auth: 
\code{\link{aemet_api_key}()}
}
\concept{aemet_auth}
