% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aemet_detect_api_key}
\alias{aemet_detect_api_key}
\title{Check if an AEMET API Key is present as environment variable}
\usage{
aemet_detect_api_key(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
The function returns TRUE/FALSE
}
\concept{helpers}
