% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_api_key.R
\name{aemet_detect_api_key}
\alias{aemet_detect_api_key}
\title{Check if an AEMET API Key is present for the current session}
\usage{
aemet_detect_api_key(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
TRUE/FALSE
}
\description{
The function would detect if an API Key is available on this session.
\itemize{
\item If an API Key is already set as an environment variable it would be
preserved
\item If no environment variable has been set and you have stored permanently
an API Key using \code{\link[=aemet_api_key]{aemet_api_key()}}, the latter would be loaded.
}
}
\examples{

aemet_detect_api_key()
}
\concept{aemet_auth}
