% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-api-key.R
\name{aemet_detect_api_key}
\alias{aemet_detect_api_key}
\alias{aemet_show_api_key}
\title{Check if an AEMET API Key is present for the current session}
\usage{
aemet_detect_api_key(...)

aemet_show_api_key(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
\code{TRUE} or \code{FALSE}. \code{aemet_show_api_key()} would display your stored API keys.
}
\description{
The function would detect if an API Key is available on this session:
\itemize{
\item If an API Key is already set as an environment variable it would be
preserved
\item If no environment variable has been set and you have stored permanently
an API Key using \code{\link[=aemet_api_key]{aemet_api_key()}}, the latter would be loaded.
}
}
\examples{

aemet_detect_api_key()

# CAUTION: This may reveal API Keys
if (FALSE) {
  aemet_show_api_key()
}
}
\seealso{
Other aemet_auth: 
\code{\link{aemet_api_key}()}
}
\concept{aemet_auth}
