% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-extremes.R
\encoding{UTF-8}
\name{aemet_extremes_clim}
\alias{aemet_extremes_clim}
\title{Extreme values for a station}
\usage{
aemet_extremes_clim(
  station = NULL,
  parameter = "T",
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{station}{A character vector of station identifiers. See
\code{\link[=aemet_stations]{aemet_stations()}}.}

\item{parameter}{A character string specifying the parameter to retrieve:
temperature (\code{"T"}), precipitation (\code{"P"}) or wind (\code{"V"}).}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{return_sf}{A logical value. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default), it
returns a \link[dplyr:tibble]{tibble}. The \CRANpkg{sf} package must be
installed.}

\item{extract_metadata}{A logical value. If \code{TRUE}, returns a
\link[dplyr:tibble]{tibble} describing the response fields. See
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{A logical value. If \code{TRUE}, displays a
\code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} unless \code{verbose = TRUE}.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object. If the function
encounters a parsing error, it returns a list.
}
\description{
Retrieves recorded extreme values for one or more stations.
}
\section{API key}{
Queries to the AEMET OpenData API require an API key. Use \code{\link[=aemet_api_key]{aemet_api_key()}}
to set it globally. Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
obs <- aemet_extremes_clim(c("9434", "3195"))
dplyr::glimpse(obs)
\dontshow{\}) # examplesIf}
}
\seealso{
Climatology:
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_monthly_clim]{aemet_monthly_clim()}},
\code{\link[=aemet_normal_clim]{aemet_normal_clim()}}
}
\concept{climatology}
