% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{aemet_extremes_clim}
\alias{aemet_extremes_clim}
\title{Extreme values for a station}
\usage{
aemet_extremes_clim(station, apikey, parameter = c("T", "P", "V"))
}
\arguments{
\item{station}{Character string as station identifier code (see \code{\link{aemet_stations}}).}

\item{apikey}{Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).}

\item{parameter}{Character string as temperature (T), precipitatation (P) or wind (W) parameter.}
}
\value{
a data.frame.
}
\description{
Get recorded extreme values for a station.
}
\examples{
\dontrun{
aemet_extremes_clim("9434", apikey, "T")
}

}
