% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{aemet_extremes_clim}
\alias{aemet_extremes_clim}
\title{Extreme values for a station}
\usage{
aemet_extremes_clim(station, apikey, parameter = c("T", "P", "V"))
}
\arguments{
\item{station}{Station identifier code (see \code{\link{aemet_stations}}).}

\item{apikey}{Personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).}

\item{parameter}{Temperature (T), Precipitatation (P) or Wind (W)}
}
\value{
a data.frame
}
\description{
Get recorded extreme values for a station
}
\examples{
\dontrun{
aemet_extremes_clim("9434", apikey, "T")
}
}
