% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_extremes.R
\name{aemet_extremes_clim}
\alias{aemet_extremes_clim}
\title{Extreme values for a station}
\usage{
aemet_extremes_clim(
  station = NULL,
  apikey = NULL,
  parameter = "T",
  verbose = FALSE,
  return_sf = FALSE
)
}
\arguments{
\item{station}{Character string with station identifier code(s)
(see \code{\link[=aemet_stations]{aemet_stations()}})}

\item{apikey}{Character string as personal API key. It can be set
on the environment variable AEMET_API_KEY, see \code{\link[=aemet_api_key]{aemet_api_key()}}.}

\item{parameter}{Character string as temperature (T),
precipitation (P) or wind (V) parameter.}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. Should the function return an \code{sf} spatial object?
If FALSE (the default value) it returns a tibble. Note that you need to
have the \code{sf} package installed.}
}
\value{
a tibble or a \code{sf} object
}
\description{
Get recorded extreme values for a station.
}
\examples{

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_extremes_clim(c("9434", "3195"))
  glimpse(obs)
}
}
\concept{aemet_api_data}
