% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-forecast-beach.R
\encoding{UTF-8}
\name{aemet_forecast_beaches}
\alias{aemet_forecast_beaches}
\title{Beach forecast dataset}
\usage{
aemet_forecast_beaches(
  x,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{x}{Character vector with beach codes to extract. See \code{\link[=aemet_beaches]{aemet_beaches()}}.}

\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default value), it
returns a \link[tibble:tbl_df]{tibble}. The \CRANpkg{sf} package must be
installed.}

\item{extract_metadata}{Logical. If \code{TRUE}, the output is a
\link[tibble:tbl_df]{tibble} with the description of the fields. See also
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{Logical. Displays a \code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} object. If
\code{verbose = TRUE}, it will not be displayed.}
}
\value{
A \link[tibble:tbl_df]{tibble} or a \CRANpkg{sf} object.
}
\description{
Get daily weather forecasts for one or more beaches. Beach codes can be
accessed with \code{\link[=aemet_beaches]{aemet_beaches()}}.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
# Forecast for beaches in Palma, Mallorca
library(dplyr)
library(ggplot2)

palma_b <- aemet_beaches() |>
  filter(ID_MUNICIPIO == "07040")

forecast_b <- aemet_forecast_beaches(palma_b$ID_PLAYA)
glimpse(forecast_b)

ggplot(forecast_b) +
  geom_line(aes(fecha, tagua_valor1, color = nombre)) +
  facet_wrap(~nombre, ncol = 1) +
  labs(
    title = "Water temperature in beaches of Palma (ES)",
    subtitle = "3-day forecast",
    x = "Date",
    y = "Temperature (Celsius)",
    color = "Beach"
  )
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aemet_beaches]{aemet_beaches()}} for beach codes.

AEMET data functions:
\code{\link[=aemet_alert_zones]{aemet_alert_zones()}},
\code{\link[=aemet_alerts]{aemet_alerts()}},
\code{\link[=aemet_beaches]{aemet_beaches()}},
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_fires]{aemet_forecast_fires()}},
\code{\link[=aemet_last_obs]{aemet_last_obs()}},
\code{\link{aemet_monthly}},
\code{\link{aemet_normal}},
\code{\link[=aemet_stations]{aemet_stations()}}

Forecast functions:
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_fires]{aemet_forecast_fires()}},
\code{\link[=aemet_forecast_tidy]{aemet_forecast_tidy()}}
}
\concept{aemet_api_data}
\concept{forecasts}
