% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-forecast-beach.R
\encoding{UTF-8}
\name{aemet_forecast_beaches}
\alias{aemet_forecast_beaches}
\title{Forecast weather at beaches}
\usage{
aemet_forecast_beaches(
  x,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{x}{A character vector of beach codes to extract. See \code{\link[=aemet_beaches]{aemet_beaches()}}.}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{return_sf}{A logical value. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default), it
returns a \link[dplyr:tibble]{tibble}. The \CRANpkg{sf} package must be
installed.}

\item{extract_metadata}{A logical value. If \code{TRUE}, the output is a
\link[dplyr:tibble]{tibble} with the description of the fields. See also
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{A logical value. If \code{TRUE}, displays a
\code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} unless \code{verbose = TRUE}.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object.
}
\description{
Retrieves daily weather forecasts for one or more beaches. Use
\code{\link[=aemet_beaches]{aemet_beaches()}} to obtain beach codes.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
# Forecast for beaches in Palma, Mallorca.
library(dplyr)
library(ggplot2)

palma_b <- aemet_beaches() |>
  filter(ID_MUNICIPIO == "07040")

forecast_b <- aemet_forecast_beaches(palma_b$ID_PLAYA)
glimpse(forecast_b)

ggplot(forecast_b) +
  geom_line(aes(fecha, tagua_valor1, color = nombre)) +
  facet_wrap(~nombre, ncol = 1) +
  labs(
    title = "Water temperature in beaches of Palma (ES)",
    subtitle = "3-day forecast",
    x = "Date",
    y = "Temperature (Celsius)",
    color = "Beach"
  )
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aemet_beaches]{aemet_beaches()}} for beach codes.

Forecasts:
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_fires]{aemet_forecast_fires()}},
\code{\link[=aemet_forecast_tidy]{aemet_forecast_tidy()}}
}
\concept{forecasts}
