% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-forecast-fires.R
\encoding{UTF-8}
\name{aemet_forecast_fires}
\alias{aemet_forecast_fires}
\title{AEMET wildfire risk forecast}
\source{
\url{https://www.aemet.es/en/eltiempo/prediccion/incendios}.
}
\usage{
aemet_forecast_fires(
  area = c("p", "c"),
  verbose = FALSE,
  extract_metadata = FALSE
)
}
\arguments{
\item{area}{A character string specifying the forecast area: \code{"p"} for
mainland Spain and the Balearic Islands or \code{"c"} for the Canary Islands.}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{extract_metadata}{A logical value. If \code{TRUE}, returns a
\link[dplyr:tibble]{tibble} describing the response fields. See
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \code{\link[terra:rast]{SpatRaster}}.
}
\description{
Retrieves daily wildfire risk levels as either tabular data or a
\code{\link[terra:rast]{SpatRaster}}.
}
\details{
The \code{SpatRaster} provides six \code{\link[=factor]{factor()}} levels: \code{"1"} for very low risk,
\code{"2"} for low risk, \code{"3"} for moderate risk, \code{"4"} for high risk, \code{"5"}
for very high risk and \code{"6"} for extreme risk.

The resulting object has several layers, each representing one of the next
seven forecast days. It also has additional attributes provided by the
\CRANpkg{terra}, including \code{\link[terra:time]{terra::time()}} and \code{\link[terra:coltab]{terra::coltab()}}.
}
\section{API key}{
Queries to the AEMET OpenData API require an API key. Use \code{\link[=aemet_api_key]{aemet_api_key()}}
to set it globally. Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
aemet_forecast_fires(extract_metadata = TRUE)

# Extract alerts.
alerts <- aemet_forecast_fires()

alerts

# Plot the raster.
library(terra)
plot(alerts, all_levels = TRUE)

# Zoom in on an area.
cyl <- mapSpain::esp_get_ccaa("Castilla y Leon", epsg = 4326)

# Convert to a SpatVector.
cyl <- vect(cyl)

fires_cyl <- crop(alerts, cyl)
title <- names(fires_cyl)[1]

plot(fires_cyl[[1]], main = title, all_levels = TRUE)
plot(cyl, add = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Forecasts:
\code{\link[=aemet_forecast_beaches]{aemet_forecast_beaches()}},
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_tidy]{aemet_forecast_tidy()}}
}
\concept{forecasts}
