% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-forecast-fires.R
\encoding{UTF-8}
\name{aemet_forecast_fires}
\alias{aemet_forecast_fires}
\title{AEMET wildfire risk forecast}
\source{
\url{https://www.aemet.es/en/eltiempo/prediccion/incendios}.
}
\usage{
aemet_forecast_fires(
  area = c("p", "c"),
  verbose = FALSE,
  extract_metadata = FALSE
)
}
\arguments{
\item{area}{Forecast area. Accepted values are:
\itemize{
\item \code{"p"} for mainland Spain and Balearic Islands.
\item \code{"c"} for Canary Islands.
}}

\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{extract_metadata}{Logical. If \code{TRUE}, the output is a
\link[tibble:tbl_df]{tibble} with the description of the fields. See also
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}
}
\value{
A \link[tibble:tbl_df]{tibble} or a \code{\link[terra:rast]{SpatRaster}}.
}
\description{
Get a \code{\link[terra:rast]{SpatRaster}} with the daily wildfire risk level.
}
\details{
The \code{SpatRaster} provides six \code{\link[=factor]{factor()}} levels with the following meaning:
\itemize{
\item \code{"1"}: Very low risk.
\item \code{"2"}: Low risk.
\item \code{"3"}: Moderate risk.
\item \code{"4"}: High risk.
\item \code{"5"}: Very high risk.
\item \code{"6"}: Extreme risk.
}

The resulting object has several layers, each one representing the forecast
for the upcoming 7 days. It also has additional attributes provided by the
\CRANpkg{terra} package, such as \code{\link[terra:time]{terra::time()}} and \code{\link[terra:coltab]{terra::coltab()}}.
}
\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
aemet_forecast_fires(extract_metadata = TRUE)

# Extract alerts.
alerts <- aemet_forecast_fires()

alerts

# Plot with terra.
library(terra)
plot(alerts, all_levels = TRUE)

# Zoom in on an area.
cyl <- mapSpain::esp_get_ccaa("Castilla y Leon", epsg = 4326)

# SpatVector
cyl <- vect(cyl)

fires_cyl <- crop(alerts, cyl)
title <- names(fires_cyl)[1]

plot(fires_cyl[[1]], main = title, all_levels = TRUE)
plot(cyl, add = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
AEMET data functions:
\code{\link[=aemet_alert_zones]{aemet_alert_zones()}},
\code{\link[=aemet_alerts]{aemet_alerts()}},
\code{\link[=aemet_beaches]{aemet_beaches()}},
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link[=aemet_forecast_beaches]{aemet_forecast_beaches()}},
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_last_obs]{aemet_last_obs()}},
\code{\link{aemet_monthly}},
\code{\link{aemet_normal}},
\code{\link[=aemet_stations]{aemet_stations()}}

Forecast functions:
\code{\link[=aemet_forecast_beaches]{aemet_forecast_beaches()}},
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_tidy]{aemet_forecast_tidy()}}
}
\concept{aemet_api_data}
\concept{forecasts}
