% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_forecast_utils.R
\name{aemet_forecast_extract}
\alias{aemet_forecast_extract}
\alias{aemet_forecast_vars_available}
\title{Helper functions for extracting forecasts}
\usage{
aemet_forecast_extract(x, var)

aemet_forecast_vars_available(x)
}
\arguments{
\item{x}{A database extracted with \code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}} or
\code{\link[=aemet_forecast_hourly]{aemet_forecast_hourly()}}.}

\item{var}{Name of the desired var to extract}
}
\value{
A vector of characters (\code{\link[=aemet_forecast_vars_available]{aemet_forecast_vars_available()}})
or a tibble (\code{\link[=aemet_forecast_extract]{aemet_forecast_extract()}}).
}
\description{
Helpers for \code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}} and \code{\link[=aemet_forecast_hourly]{aemet_forecast_hourly()}}:
\itemize{
\item \code{\link[=aemet_forecast_vars_available]{aemet_forecast_vars_available()}} extracts the values available on
the dataset.
\item \code{\link[=aemet_forecast_extract]{aemet_forecast_extract()}} produces a \code{tibble} with the forecast
for \code{var}.
}
}
\examples{
\dontshow{if (aemet_detect_api_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Hourly values
hourly <- aemet_forecast_hourly(c("15030", "38038"))

# Vars available
aemet_forecast_vars_available(hourly)

# Get temperature
temp <- aemet_forecast_extract(hourly, "temperatura")

library(dplyr)
# Make hour: The approach differs depending on the metric
temp_end <- temp \%>\%
  mutate(
    hour = paste0(temperatura_periodo, ":00"),
    forecast_time = as.POSIXct(paste(fecha, hour, tz = "Europe/Madrid"))
  )


# Plot

library(ggplot2)

ggplot(temp_end) +
  geom_col(aes(forecast_time, temperatura_value), fill = "blue4") +
  facet_wrap(~nombre, nrow = 2) +
  scale_x_datetime(labels = scales::label_date_short()) +
  scale_y_continuous(labels = scales::label_number(suffix = "º")) +
  labs(
    x = "", y = "",
    title = "Forecast: Temperature",
    subtitle = paste("Forecast produced on", format(temp_end$elaborado[1], usetz = TRUE))
  )
\dontshow{\}) # examplesIf}
}
\seealso{
Other forecasts: 
\code{\link{aemet_forecast_daily}()}
}
\concept{forecasts}
