% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-last-obs.R
\encoding{UTF-8}
\name{aemet_last_obs}
\alias{aemet_last_obs}
\title{Latest observations from weather stations}
\usage{
aemet_last_obs(
  station = "all",
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{station}{Character string with station identifier code(s) (see
\code{\link[=aemet_stations]{aemet_stations()}}) or \code{"all"} for all the stations.}

\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default value), it
returns a \link[dplyr:tibble]{tibble}. The \CRANpkg{sf} package must be
installed.}

\item{extract_metadata}{Logical. If \code{TRUE}, the output is a
\link[dplyr:tibble]{tibble} with the description of the fields. See also
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{Logical. Displays a \code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} object. If
\code{verbose = TRUE}, it will not be displayed.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object.
}
\description{
Retrieves the latest observations for one or more weather stations.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
obs <- aemet_last_obs(c("9434", "3195"))
dplyr::glimpse(obs)
\dontshow{\}) # examplesIf}
}
\seealso{
AEMET data functions:
\code{\link[=aemet_alert_zones]{aemet_alert_zones()}},
\code{\link[=aemet_alerts]{aemet_alerts()}},
\code{\link[=aemet_beaches]{aemet_beaches()}},
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link[=aemet_forecast_beaches]{aemet_forecast_beaches()}},
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_fires]{aemet_forecast_fires()}},
\code{\link{aemet_monthly}},
\code{\link{aemet_normal}},
\code{\link[=aemet_stations]{aemet_stations()}}
}
\concept{aemet_api_data}
