% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{aemet_last_obs}
\alias{aemet_last_obs}
\title{Last observation values for a station}
\usage{
aemet_last_obs(station, apikey)
}
\arguments{
\item{station}{Character string as station identifier code (see \code{\link{aemet_stations}}).}

\item{apikey}{Character string as personal API key (see \url{https://opendata.aemet.es/centrodedescargas/obtencionAPIKey}).}
}
\value{
a data.frame.
}
\description{
Get last observation values for a station.
}
\examples{
\dontrun{
aemet_last_obs("9434", apikey)
}

}
