% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_observacion.R
\name{aemet_last_obs}
\alias{aemet_last_obs}
\title{Last observation values for a station}
\usage{
aemet_last_obs(station = "all", apikey = NULL, verbose = FALSE)
}
\arguments{
\item{station}{Character string with station identifier code(s)
(see \code{\link[=aemet_stations]{aemet_stations()}}) or "all" for all the values.}

\item{apikey}{Character string as personal API key. It can be set
on the environment variable AEMET_API_KEY, see \code{\link[=aemet_api_key]{aemet_api_key()}}.}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}
}
\value{
a tibble.
}
\description{
Get last observation values for a station.
}
\examples{

# Run this example only if AEMET_API_KEY is set
apikey <- Sys.getenv("AEMET_API_KEY")
if (apikey != "") {
  library(tibble)
  obs <- aemet_last_obs()
  glimpse(obs)
}
}
\concept{aemet_observaciones}
