% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-monthly.R
\encoding{UTF-8}
\name{aemet_monthly}
\alias{aemet_monthly}
\alias{aemet_monthly_clim}
\alias{aemet_monthly_period}
\alias{aemet_monthly_period_all}
\title{Monthly/annual climatology values}
\usage{
aemet_monthly_clim(
  station = NULL,
  year = as.integer(format(Sys.Date(), "\%Y")),
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_monthly_period(
  station = NULL,
  start = as.integer(format(Sys.Date(), "\%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_monthly_period_all(
  start = as.integer(format(Sys.Date(), "\%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{station}{Character string with station identifier code(s). See
\code{\link[=aemet_stations]{aemet_stations()}}.}

\item{year}{Numeric value with year (format: \code{YYYY}).}

\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default value), it
returns a \link[dplyr:tibble]{tibble}. The \CRANpkg{sf} package must be
installed.}

\item{extract_metadata}{Logical. If \code{TRUE}, the output is a
\link[dplyr:tibble]{tibble} with the description of the fields. See also
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{Logical. Displays a \code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} object. If
\code{verbose = TRUE}, it will not be displayed.}

\item{start}{Numeric value with the start year (format: \code{YYYY}).}

\item{end}{Numeric value with the end year (format: \code{YYYY}).}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object.
}
\description{
Get monthly/annual climatology values for one or more stations.
\code{aemet_monthly_period()} and \code{aemet_monthly_period_all()} allow requests
that span several years.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
obs <- aemet_monthly_clim(station = c("9434", "3195"), year = 2000)
dplyr::glimpse(obs)
\dontshow{\}) # examplesIf}
}
\seealso{
AEMET data functions:
\code{\link[=aemet_alert_zones]{aemet_alert_zones()}},
\code{\link[=aemet_alerts]{aemet_alerts()}},
\code{\link[=aemet_beaches]{aemet_beaches()}},
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link[=aemet_forecast_beaches]{aemet_forecast_beaches()}},
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_fires]{aemet_forecast_fires()}},
\code{\link[=aemet_last_obs]{aemet_last_obs()}},
\code{\link{aemet_normal}},
\code{\link[=aemet_stations]{aemet_stations()}}
}
\concept{aemet_api_data}
