% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_monthly.R
\name{aemet_monthly_clim}
\alias{aemet_monthly_clim}
\alias{aemet_monthly_period}
\alias{aemet_monthly_period_all}
\title{Monthly/annual climatology}
\usage{
aemet_monthly_clim(
  station = NULL,
  apikey = NULL,
  year = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_monthly_period(
  station = NULL,
  apikey = NULL,
  start = 2018,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_monthly_period_all(
  apikey = NULL,
  start = 2019,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)
}
\arguments{
\item{station}{Character string with station identifier code(s)
(see \code{\link[=aemet_stations]{aemet_stations()}})}

\item{apikey}{Character string as personal API key. It can be set
on the environment variable AEMET_API_KEY, see \code{\link[=aemet_api_key]{aemet_api_key()}}.}

\item{year}{Numeric value as date (format: \%Y).}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. Should the function return an \code{sf} spatial object?
If FALSE (the default value) it returns a tibble. Note that you need to
have the \code{sf} package installed.}

\item{start}{Numeric value as start year (format: \%Y).}

\item{end}{Numeric value as end year (format: \%Y).}
}
\value{
a tibble or a \code{sf} object
}
\description{
Get monthly/annual climatology values for a station or all the stations.
\code{aemet_monthly_period()} and \code{aemet_monthly_period_all()} allows requests
that span several years.
}
\examples{

# Run this example only if AEMET_API_KEY is set
apikey <- Sys.getenv("AEMET_API_KEY")
if (apikey != "") {
  library(tibble)
  obs <- aemet_monthly_clim(station = c("9434", "3195"), year = 2000)
  glimpse(obs)
}
}
\concept{aemet_valores}
