% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{aemet_munic}
\alias{aemet_munic}
\title{Municipalities of Spain}
\format{
A \link[dplyr:tibble]{tibble} with
8,132 rows and fields:
\describe{
\item{municipio}{INE code of the municipality.}
\item{municipio_nombre}{INE name of the municipality.}
\item{cpro}{INE code of the province.}
\item{cpro_nombre}{INE name of the province.}
\item{codauto}{INE code of the autonomous community.}
\item{codauto_nombre}{INE name of the autonomous community.}
}
}
\source{
INE municipality codes by province:
\url{https://www.ine.es/daco/daco42/codmun/diccionario25.xlsx}.
}
\description{
A \link[dplyr:tibble]{tibble} containing all municipalities of Spain as defined
by the INE (Instituto Nacional de Estadística) as of January 2025.
}
\examples{

data(aemet_munic)

aemet_munic
}
\seealso{
\itemize{
\item \code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}} retrieves daily municipality forecasts.
\item \code{\link[=aemet_forecast_hourly]{aemet_forecast_hourly()}} retrieves hourly municipality forecasts.
}

Included datasets:
\code{\link{climaemet_9434_climatogram}},
\code{\link{climaemet_9434_temp}},
\code{\link{climaemet_9434_wind}}
}
\concept{datasets}
\keyword{datasets}
