% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{aemet_munic}
\alias{aemet_munic}
\title{Data set with all the municipalities of Spain}
\format{
A \code{\link[tibble:tibble]{tibble}} with
8,131 rows and fields:
\describe{
\item{municipio}{INE code of the municipality.}
\item{municipio_nombre}{INE name of the municipality.}
\item{cpro}{INE code of the province.}
\item{cpro_nombre}{INE name of the province.}
\item{codauto}{INE code of the autonomous community.}
\item{codauto_nombre}{INE code of the autonomous community.}
}
}
\source{
INE, \href{https://www.ine.es/en/daco/daco42/codmun/codmunmapa_en.htm}{Municipality codes by province}
}
\description{
A \code{\link[tibble:tibble]{tibble}} with all the municipalities of  Spain as
defined by the INE (Instituto Nacional de Estadistica) as of January 2020.
}
\examples{

data(aemet_munic)

aemet_munic
}
\seealso{
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_hourly]{aemet_forecast_hourly()}}

Other dataset: 
\code{\link{climaemet_9434_climatogram}},
\code{\link{climaemet_9434_temp}},
\code{\link{climaemet_9434_wind}}
}
\concept{dataset}
\concept{forecast}
