% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{aemet_munic}
\alias{aemet_munic}
\title{Data set with all the municipalities of Spain}
\format{
A \link[tibble:tbl_df-class]{tibble} with
8,132 rows and fields:
\describe{
\item{municipio}{INE code of the municipality.}
\item{municipio_nombre}{INE name of the municipality.}
\item{cpro}{INE code of the province.}
\item{cpro_nombre}{INE name of the province.}
\item{codauto}{INE code of the autonomous community.}
\item{codauto_nombre}{INE code of the autonomous community.}
}
}
\source{
INE,Municipality codes by province:

\url{https://www.ine.es/daco/daco42/codmun/diccionario25.xlsx}
}
\description{
A \link[tibble:tbl_df-class]{tibble} with all the municipalities of Spain as
defined by the INE (Instituto Nacional de Estadistica) as of January 2025.
}
\examples{

data(aemet_munic)

aemet_munic
}
\seealso{
\code{\link[=aemet_forecast_daily]{aemet_forecast_daily()}},
\code{\link[=aemet_forecast_hourly]{aemet_forecast_hourly()}}

Other dataset: 
\code{\link{climaemet_9434_climatogram}},
\code{\link{climaemet_9434_temp}},
\code{\link{climaemet_9434_wind}}
}
\concept{dataset}
\concept{forecast}
