% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-normal.R
\encoding{UTF-8}
\name{aemet_normal_clim}
\alias{aemet_normal_clim}
\alias{aemet_normal}
\alias{aemet_normal_clim_all}
\title{Climatological normal values}
\usage{
aemet_normal_clim(
  station = NULL,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_normal_clim_all(
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{station}{A character vector of station identifiers (see
\code{\link[=aemet_stations]{aemet_stations()}}) or \code{"all"} for all stations.}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{return_sf}{A logical value. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default), it
returns a \link[dplyr:tibble]{tibble}. \CRANpkg{sf} must be installed.}

\item{extract_metadata}{A logical value. If \code{TRUE}, returns a
\link[dplyr:tibble]{tibble} describing the response fields. See
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{A logical value. If \code{TRUE}, displays a
\code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} unless \code{verbose = TRUE}.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object.
}
\description{
Retrieves climatological normal values for a station or for all stations
with \code{\link[=aemet_normal_clim_all]{aemet_normal_clim_all()}}. The standard normal period is 1981–2010.
}
\note{
Code adapted from the project at \url{https://github.com/SevillaR/aemet}.
}
\section{API key}{
Queries to the AEMET OpenData API require an API key. Use \code{\link[=aemet_api_key]{aemet_api_key()}}
to set it globally. Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
obs <- aemet_normal_clim(c("9434", "3195"))
dplyr::glimpse(obs)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aemet_stations]{aemet_stations()}} for station identifiers.

Climatology:
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link[=aemet_monthly_clim]{aemet_monthly_clim()}}
}
\concept{climatology}
