% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-normal.R
\encoding{UTF-8}
\name{aemet_normal}
\alias{aemet_normal}
\alias{aemet_normal_clim}
\alias{aemet_normal_clim_all}
\title{Climatological normal values}
\usage{
aemet_normal_clim(
  station = NULL,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_normal_clim_all(
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)
}
\arguments{
\item{station}{Character string with station identifier code(s) (see
\code{\link[=aemet_stations]{aemet_stations()}}) or \code{"all"} for all the stations.}

\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default value), it
returns a \link[dplyr:tibble]{tibble}. The \CRANpkg{sf} package must be
installed.}

\item{extract_metadata}{Logical. If \code{TRUE}, the output is a
\link[dplyr:tibble]{tibble} with the description of the fields. See also
\code{\link[=get_metadata_aemet]{get_metadata_aemet()}}.}

\item{progress}{Logical. Displays a \code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}} object. If
\code{verbose = TRUE}, it will not be displayed.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object.
}
\description{
Retrieves climatological normal values for a station or for all stations
with \code{aemet_normal_clim_all()}. The standard normal period is 1981-2010.
}
\note{
Code modified from project \url{https://github.com/SevillaR/aemet}.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
obs <- aemet_normal_clim(c("9434", "3195"))
dplyr::glimpse(obs)
\dontshow{\}) # examplesIf}
}
\seealso{
Climatology:
\code{\link[=aemet_daily_clim]{aemet_daily_clim()}},
\code{\link[=aemet_extremes_clim]{aemet_extremes_clim()}},
\code{\link{aemet_monthly}}
}
\concept{climatology}
