% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-stations.R
\encoding{UTF-8}
\name{aemet_stations}
\alias{aemet_stations}
\title{AEMET stations}
\usage{
aemet_stations(verbose = FALSE, return_sf = FALSE)
}
\arguments{
\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{return_sf}{A logical value. If \code{TRUE}, the function returns an
\code{\link[sf:st_sf]{sf}} spatial object. If \code{FALSE} (the default), it
returns a \link[dplyr:tibble]{tibble}. The \CRANpkg{sf} package must be
installed.}
}
\value{
A \link[dplyr:tibble]{tibble} or a \CRANpkg{sf} object.
}
\description{
Retrieves the weather stations available from the AEMET API.
}
\note{
Code modified from project \url{https://github.com/SevillaR/aemet}.
}
\section{Caching}{

The first result retrieved in each session is temporarily cached in
\code{\link[=tempdir]{tempdir()}} to avoid unnecessary requests.
}

\section{API key}{
Queries to the AEMET OpenData API require an API key. Use \code{\link[=aemet_api_key]{aemet_api_key()}}
to set it globally. Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
library(dplyr)
stations <- aemet_stations()
stations

# Cached during this R session.
stations2 <- aemet_stations(verbose = TRUE)

identical(stations, stations2)
\dontshow{\}) # examplesIf}
}
\concept{locations}
