% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{aemet_stations}
\alias{aemet_stations}
\title{AEMET stations}
\usage{
aemet_stations(apikey)
}
\arguments{
\item{apikey}{Character string as API key (https://opendata.aemet.es/centrodedescargas/obtencionAPIKey).}
}
\value{
a data.frame.
}
\description{
Get AEMET stations.
}
\note{
Code modified from project https://github.com/SevillaR/aemet
}
\examples{
\dontrun{
stations <- aemet_stations(apikey)
}

}
