% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_valores.R
\name{aemet_stations}
\alias{aemet_stations}
\title{AEMET stations}
\usage{
aemet_stations(apikey = NULL, verbose = FALSE)
}
\arguments{
\item{apikey}{Character string as personal API key. It can be set
on the environment variable AEMET_API_KEY, see \code{\link[=aemet_api_key]{aemet_api_key()}}.}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}
}
\value{
A tibble
}
\description{
Get AEMET stations.
}
\note{
Code modified from project \url{https://github.com/SevillaR/aemet}
}
\examples{
# Run this example only if AEMET_API_KEY is set
apikey <- Sys.getenv("AEMET_API_KEY")
if (apikey != "") {
  library(tibble)
  stations <- aemet_stations(apikey)
  glimpse(stations)
}
}
\concept{aemet_valores}
