% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_stations.R
\name{aemet_stations}
\alias{aemet_stations}
\title{AEMET stations}
\usage{
aemet_stations(verbose = FALSE, return_sf = FALSE)
}
\arguments{
\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical. Should the function return an \code{sf} spatial object?
If FALSE (the default value) it returns a tibble. Note that you need to
have the \code{sf} package installed.}
}
\value{
a tibble or a \code{sf} object
}
\description{
Get AEMET stations.
}
\note{
Code modified from project \url{https://github.com/SevillaR/aemet}
}
\examples{
# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  library(tibble)
  stations <- aemet_stations()
  stations
}
}
\concept{aemet_api_data}
