% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_stations.R
\name{aemet_stations}
\alias{aemet_stations}
\title{AEMET stations}
\usage{
aemet_stations(verbose = FALSE, return_sf = FALSE)
}
\arguments{
\item{verbose}{Logical \code{TRUE/FALSE}. Provides information about the flow of
information between the client and server.}

\item{return_sf}{Logical \code{TRUE} or \code{FALSE}.
Should the function return an \code{sf} spatial object? If \code{FALSE} (the default
value) it returns a \link[tibble:tibble]{tibble}. Note that you need to have
the \CRANpkg{sf} package installed.}
}
\value{
A \link[tibble:tibble]{tibble} or a \CRANpkg{sf} object
}
\description{
Get AEMET stations.
}
\details{
The first result of the API call on each session is (temporarily) cached in
the assigned \code{\link[=tempdir]{tempdir()}} for avoiding unneeded API calls.
}
\note{
Code modified from project \url{https://github.com/SevillaR/aemet}
}
\section{API Key}{
You need to set your API Key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
}

\examples{
\dontshow{if (aemet_detect_api_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(tibble)
stations <- aemet_stations()
stations

# Cached during this R session
stations2 <- aemet_stations(verbose = TRUE)

identical(stations, stations2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other aemet_api_data: 
\code{\link{aemet_daily_clim}()},
\code{\link{aemet_extremes_clim}()},
\code{\link{aemet_forecast_daily}()},
\code{\link{aemet_last_obs}()},
\code{\link{aemet_monthly}},
\code{\link{aemet_normal}}
}
\concept{aemet_api_data}
