% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climaemet-package.R
\docType{package}
\name{climaemet-package}
\alias{climaemet}
\alias{climaemet-package}
\title{climaemet: Tools for AEMET Climate Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to download climate data from the Spanish Meteorological Agency (AEMET) directly in R using the AEMET API and to create scientific visualizations, including climate charts, trend analyses of climate time series, temperature and precipitation anomaly maps, warming stripes and climatograms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/climaemet/}
  \item \url{https://github.com/rOpenSpain/climaemet}
  \item Report bugs at \url{https://github.com/rOpenSpain/climaemet/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})
  \item Manuel Pizarro (\href{https://orcid.org/0000-0002-6981-0154}{ORCID}) [copyright holder]
  \item Gema Fernández-Avilés (\href{https://orcid.org/0000-0001-5934-1916}{ORCID})
}

Other contributors:
\itemize{
  \item AEMET (\href{https://ror.org/04kxf1r09}{ROR}) [copyright holder]
}

}
\keyword{internal}
