% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climaemet-package.R
\docType{package}
\name{climaemet-package}
\alias{climaemet}
\alias{climaemet-package}
\title{climaemet: R Climate AEMET Tools}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

R interface to download the climatic data of the Spanish Meteorological Agency (AEMET) directly from R using their API (https://opendata.aemet.es/) and create scientific graphs.
}
\keyword{internal}
