% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climaemet-package.R
\docType{package}
\name{climaemet-package}
\alias{climaemet}
\alias{climaemet-package}
\title{climaemet: Climate AEMET Tools}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tools to download the climatic data of the Spanish Meteorological Agency (AEMET) directly from R using their API <https://opendata.aemet.es/> and create scientific graphs (climate charts, trend analysis of climate time series, temperature and precipitation anomalies maps, warming stripes graphics, climatograms, etc.).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mpizarrotig.github.io/climaemet}
  \item \url{https://github.com/mpizarrotig/climaemet}
  \item Report bugs at \url{https://github.com/mpizarrotig/climaemet/issues}
}

}
\keyword{internal}
