% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{climaemet_9434_climatogram}
\alias{climaemet_9434_climatogram}
\title{Climatological normals for Zaragoza Airport ("9434"), 1981–2010}
\format{
A data frame with four rows and 12 columns. Columns \code{1} through \code{12}
represent months from January through December. Rows contain:
\itemize{
\item \code{p_mes_md}: precipitation (mm).
\item \code{tm_max_md}: maximum temperature (Celsius).
\item \code{tm_min_md}: minimum temperature (Celsius).
\item \code{ta_min_min}: absolute monthly minimum temperature (Celsius).
}
}
\source{
AEMET.
}
\description{
Climatological normal data for Zaragoza Airport (1981–2010). This example
dataset is used to create Walter-Lieth climate diagrams.
}
\examples{
data(climaemet_9434_climatogram)
}
\seealso{
\itemize{
\item \code{\link[=climatogram_normal]{climatogram_normal()}} and \code{\link[=climatogram_period]{climatogram_period()}} retrieve climatology
data.
\item \code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}} creates Walter-Lieth climate diagrams.
}
}
\concept{datasets}
\keyword{datasets}
