% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{climaemet_9434_climatogram}
\alias{climaemet_9434_climatogram}
\title{Climatogram data for Zaragoza Airport ("9434"), 1981-2010}
\format{
A \link{data.frame} with 12 month columns and four rows:
\describe{
\item{p_mes_md}{Precipitation (mm).}
\item{tm_max_md}{Maximum temperature (Celsius).}
\item{tm_min_md}{Minimum temperature (Celsius).}
\item{ta_min_md}{Absolute monthly minimum temperature (Celsius).}
}
}
\source{
AEMET.
}
\description{
Climatological normal data for Zaragoza Airport (1981-2010). This example
dataset is used to plot climatograms.
}
\examples{
data(climaemet_9434_climatogram)
}
\seealso{
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}, \code{\link[=climatogram_period]{climatogram_period()}},
\code{\link[=climatogram_normal]{climatogram_normal()}}

Included datasets:
\code{\link{aemet_munic}},
\code{\link{climaemet_9434_temp}},
\code{\link{climaemet_9434_wind}}

Climatogram functions:
\code{\link[=climatogram_normal]{climatogram_normal()}},
\code{\link[=climatogram_period]{climatogram_period()}},
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}
}
\concept{climatogram}
\concept{dataset}
