% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{climaemet_9434_climatogram}
\alias{climaemet_9434_climatogram}
\title{Climatological normals for Zaragoza Airport ("9434"), 1981–2010}
\format{
A data frame with 12 monthly columns and four rows:
\describe{
\item{p_mes_md}{Precipitation (mm).}
\item{tm_max_md}{Maximum temperature (Celsius).}
\item{tm_min_md}{Minimum temperature (Celsius).}
\item{ta_min_md}{Absolute monthly minimum temperature (Celsius).}
}
}
\source{
AEMET.
}
\description{
Climatological normal data for Zaragoza Airport (1981–2010). This example
dataset is used to create Walter-Lieth climate diagrams.
}
\examples{
data(climaemet_9434_climatogram)
}
\seealso{
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}, \code{\link[=climatogram_period]{climatogram_period()}},
\code{\link[=climatogram_normal]{climatogram_normal()}}.
}
\concept{datasets}
