% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climaemet_9434_climatogram}
\alias{climaemet_9434_climatogram}
\title{Climatogram data for Zaragoza Airport ("9434") period 1981-2010}
\format{
A data.frame with columns 1 to 12 (months) and rows:
\itemize{
\item \strong{p_mes_md}: Precipitation (mm).
\item \strong{tm_max_md}: Maximum temperature (Celsius).
\item \strong{tm_min_md}: Minimum temperature (Celsius).
\item \strong{ta_min_md}: Absolute monthly minimum temperature (Celsius).
}
}
\source{
AEMET.
}
\description{
Normal data for Zaragoza Airport (1981-2010). This is an example dataset
used to plot climatograms.
}
\examples{
data(climaemet_9434_climatogram)
}
\seealso{
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}, \code{\link[=climatogram_period]{climatogram_period()}},
\code{\link[=climatogram_normal]{climatogram_normal()}}

Other dataset: 
\code{\link{climaemet_9434_temp}},
\code{\link{climaemet_9434_wind}}
}
\concept{dataset}
