% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{climaemet_9434_temp}
\alias{climaemet_9434_temp}
\title{Average annual temperatures for Zaragoza Airport ("9434"), 1950-2020}
\format{
A \link[dplyr:tibble]{tibble} with columns:
\describe{
\item{year}{Year of reference.}
\item{indicativo}{Identifier of the station.}
\item{temp}{Average temperature (Celsius).}
}
}
\source{
AEMET.
}
\description{
Yearly observations of average temperature for Zaragoza Airport
(1950-2020). This is an example dataset.
}
\examples{
data(climaemet_9434_temp)
}
\seealso{
Included datasets:
\code{\link{aemet_munic}},
\code{\link{climaemet_9434_climatogram}},
\code{\link{climaemet_9434_wind}}

Warming stripes functions:
\code{\link[=climatestripes_station]{climatestripes_station()}},
\code{\link[=ggstripes]{ggstripes()}}
}
\concept{dataset}
\concept{stripes}
