% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climaemet_9434_temp}
\alias{climaemet_9434_temp}
\title{Average annual temperatures for Zaragoza Airport ("9434") period 1950-2020}
\format{
A tibble with columns:
\itemize{
\item \strong{year}: Year of reference.
\item \strong{indicativo}: Identifier of the station.
\item \strong{temp}: Avg temperature (Celsius).
}
}
\source{
AEMET.
}
\description{
Yearly observations of average temperature for Zaragoza Airport
(1950-2020). This is an example dataset.
}
\examples{
data(climaemet_9434_temp)
}
\seealso{
Other dataset: 
\code{\link{aemet_munic}},
\code{\link{climaemet_9434_climatogram}},
\code{\link{climaemet_9434_wind}}

Other stripes: 
\code{\link{climatestripes_station}()},
\code{\link{ggstripes}()}
}
\concept{dataset}
\concept{stripes}
