% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{climaemet_9434_wind}
\alias{climaemet_9434_wind}
\title{Wind conditions for Zaragoza Airport ("9434"), 2000-2020}
\format{
A \link[dplyr:tibble]{tibble} with columns:

\describe{
\item{fecha}{Date of observation.}
\item{dir}{Wind direction (0-360 degrees).}
\item{velmedia}{Average wind speed (km/h).}
}
}
\source{
AEMET.
}
\description{
Daily observations of wind speed and direction for Zaragoza Airport
(2000-2020). This is an example dataset.
}
\examples{
data(climaemet_9434_wind)
}
\seealso{
\code{\link[=windrose_days]{windrose_days()}}, \code{\link[=windrose_period]{windrose_period()}}, \code{\link[=ggwindrose]{ggwindrose()}}.

Other dataset:
\code{\link{aemet_munic}},
\code{\link{climaemet_9434_climatogram}},
\code{\link{climaemet_9434_temp}}
}
\concept{dataset}
