% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climaemet_9434_wind}
\alias{climaemet_9434_wind}
\title{Wind conditions for Zaragoza Airport ("9434") period 2000-2020}
\format{
A \link[tibble:tbl_df-class]{tibble} with columns:

\describe{
\item{fecha}{Date of observation.}
\item{dir}{Wind directions (0-360).}
\item{velmedia}{Average wind speed (km/h)}
}
}
\source{
AEMET.
}
\description{
Daily observations of wind speed and directions for Zaragoza Airport
(2000-2020). This is an example dataset.
}
\examples{
data(climaemet_9434_wind)
}
\seealso{
Other dataset: 
\code{\link{aemet_munic}},
\code{\link{climaemet_9434_climatogram}},
\code{\link{climaemet_9434_temp}}

Other wind: 
\code{\link{ggwindrose}()},
\code{\link{windrose_days}()},
\code{\link{windrose_period}()}
}
\concept{dataset}
\concept{wind}
