% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climaemet_9434_wind}
\alias{climaemet_9434_wind}
\title{Wind conditions for Zaragoza Airport ("9434") period 2000-2020}
\format{
A tibble with columns:
\itemize{
\item \strong{fecha}: Date of observation.
\item \strong{dir}: Wind directions (0-360).
\item \strong{velmedia}: Avg wind speed (km/h).
}
}
\source{
AEMET.
}
\description{
Daily observations of wind speed and directions for Zaragoza Airport
(2000-2020). This is an example dataset.
}
\examples{
data(climaemet_9434_wind)
}
\seealso{
Other dataset: 
\code{\link{climaemet_9434_climatogram}},
\code{\link{climaemet_9434_temp}}
}
\concept{dataset}
