% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climaemet-news.R
\name{climaemet_news}
\alias{climaemet_news}
\title{Show the latest news of \CRANpkg{climaemet}}
\usage{
climaemet_news()
}
\value{
Nothing, this function is called by its side effects.
}
\description{
Open the \code{NEWS} file of the \CRANpkg{climaemet} package.
}
\examples{
\dontrun{
climaemet_news()
}

}
\seealso{
Other helpers: 
\code{\link{dms2decdegrees}()},
\code{\link{first_day_of_year}()}
}
\concept{helpers}
\keyword{internal}
