% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climaemet_news.R
\name{climaemet_news}
\alias{climaemet_news}
\title{climaemet_news}
\usage{
climaemet_news()
}
\description{
Show the NEWS file of the \pkg{climaemet} package.
}
\details{
(See description)
}
\concept{helpers}
