% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climaemet_news.R
\name{climaemet_news}
\alias{climaemet_news}
\title{climaemet_news}
\usage{
climaemet_news()
}
\value{
Open NEWS from \code{climaemet}.
}
\description{
Show the NEWS file of the \CRANpkg{climaemet} package.
}
\details{
(See description)
}
\examples{
\dontrun{
climaemet_news()
}

}
\seealso{
Other helpers: 
\code{\link{dms2decdegrees}()},
\code{\link{first_day_of_year}()}
}
\concept{helpers}
