% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripes.R
\encoding{UTF-8}
\name{climatestripes_station}
\alias{climatestripes_station}
\alias{ggstripes_station}
\title{Warming stripes for a weather station}
\usage{
climatestripes_station(
  station,
  start = 1950,
  end = 2020,
  with_labels = "yes",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{station}{A character vector of station identifiers. See
\code{\link[=aemet_stations]{aemet_stations()}}.}

\item{start}{A numeric value specifying the start year in \code{YYYY} format.}

\item{end}{A numeric value specifying the end year in \code{YYYY} format.}

\item{with_labels}{A character string indicating whether to display plot
labels, either \code{"yes"} or \code{"no"}.}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{...}{
  Arguments passed on to \code{\link{ggstripes}}
  \describe{
    \item{\code{n_temp}}{The number of colors in the palette. Defaults to \code{11}.}
    \item{\code{col_pal}}{A character string specifying the
\code{\link[=hcl.pals]{hcl.pals()}} color palette to be used for plotting.}
  }}
}
\value{
A \CRANpkg{ggplot2} object. See \code{help("ggplot2")}.
}
\description{
Plots warming stripes for a weather station over a specified period.
}
\note{
Professor Ed Hawkins of the University of Reading developed the "warming
stripes" concept to communicate climate change risks as simply as possible.
For more details, see
\href{https://showyourstripes.info/}{ShowYourStripes}.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
\donttest{

# Do not run this example.
if (FALSE) {
  # Downloading data may take a few minutes.
  climatestripes_station(
    "9434",
    start = 2020,
    end = 2024,
    with_labels = "yes",
    col_pal = "Inferno"
  )
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
\link{climaemet_9434_wind}.

Warming stripes:
\code{\link[=ggstripes]{ggstripes()}}
}
\concept{stripes}
