% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripes.R
\name{climatestripes_station}
\alias{climatestripes_station}
\title{Station climate stripes graph}
\usage{
climatestripes_station(
  station,
  start = 1950,
  end = 2020,
  with_labels = "yes",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{station}{Character string as station identifier code
(see \code{\link[=aemet_stations]{aemet_stations()}}).}

\item{start}{Numeric value as start year (format: \%Y).}

\item{end}{Numeric value as end year (format: \%Y).}

\item{with_labels}{Character string as yes/no. Indicates whether to use
labels for the graph or not.}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}

\item{...}{Further arguments of \code{\link[=ggstripes]{ggstripes()}}}
}
\value{
a \code{ggplot2} object
}
\description{
Plot climate stripes graph for a station
}
\examples{

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  climatestripes_station(
    "9434",
    start = 2010,
    end = 2020,
    with_labels = "yes",
    col_pal = "Inferno"
  )
}
}
\seealso{
\code{\link[=ggstripes]{ggstripes()}}
}
\concept{aemet_plots}
