% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climaemet_main.R
\name{climatestripes_station}
\alias{climatestripes_station}
\title{Station climate stripes graph}
\usage{
climatestripes_station(
  station,
  apikey = NULL,
  start = 1950,
  end = 2020,
  with_labels = "yes",
  verbose = FALSE
)
}
\arguments{
\item{station}{Character string as station identifier code
(see \code{\link[=aemet_stations]{aemet_stations()}}).}

\item{apikey}{Character string as personal API key. It can be set
on the environment variable AEMET_API_KEY, see \code{\link[=aemet_api_key]{aemet_api_key()}}.}

\item{start}{Numeric value as start year (format: \%Y).}

\item{end}{Numeric value as end year (format: \%Y).}

\item{with_labels}{Character string as yes/no. Indicates whether to use
labels for the graph or not.}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}
}
\value{
a plot.
}
\description{
Plot climate stripes graph for a station
}
\examples{

# Run this example only if AEMET_API_KEY is set
apikey <- Sys.getenv("AEMET_API_KEY")
if (apikey != "") {
  climatestripes_station("9434",
    start = 2010, end = 2020, with_labels = "yes"
  )
}
}
\concept{aemet_plots}
