% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatogram.R
\encoding{UTF-8}
\name{climatogram_normal}
\alias{climatogram_normal}
\title{Walter-Lieth climate diagram from climatological normals}
\usage{
climatogram_normal(
  station,
  labels = "en",
  verbose = FALSE,
  ggplot2 = TRUE,
  ...
)
}
\arguments{
\item{station}{A character vector of station identifiers. See
\code{\link[=aemet_stations]{aemet_stations()}}.}

\item{labels}{A character string specifying the language for the x-axis
month labels, such as \code{"en"} (English), \code{"es"} (Spanish) or \code{"fr"}
(French).}

\item{verbose}{A logical value. If \code{TRUE}, displays information about the
exchange between the client and server.}

\item{ggplot2}{A logical value. If \code{TRUE}, the function uses
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}. If \code{FALSE}, it uses \code{\link[climatol:diagwl]{climatol::diagwl()}}.}

\item{...}{Further arguments passed to
\code{\link[climatol:diagwl]{climatol::diagwl()}} or \code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}, depending on the
value of \code{ggplot2}.}
}
\value{
A plot.
}
\description{
Plots a Walter-Lieth climate diagram from climatological normal values for a
station. The diagram summarizes local climate conditions for 1981–2010.
}
\note{
The implementation is based on \CRANpkg{climatol}.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
climatogram_normal("9434")
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Walter, H. K., Harnickell, E., Lieth, F. H. H. and Rehder, H. (1967).
\emph{Klimadiagramm-weltatlas}. Jena: Fischer.
\item Guijarro J. A. (2023).
\emph{climatol: Climate Tools (Series Homogenization and Derived Products)}. R
package version 4.0.0, \url{https://climatol.eu}.
}
}
\seealso{
\link{climaemet_9434_climatogram}.

Walter-Lieth climate diagrams:
\code{\link[=climatogram_period]{climatogram_period()}},
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}
}
\concept{climatogram}
