% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatogram.R
\encoding{UTF-8}
\name{climatogram_period}
\alias{climatogram_period}
\title{Walter & Lieth climatic diagram for a time period}
\usage{
climatogram_period(
  station = NULL,
  start = 1990,
  end = 2020,
  labels = "en",
  verbose = FALSE,
  ggplot2 = TRUE,
  ...
)
}
\arguments{
\item{station}{Character string with station identifier code(s). See
\code{\link[=aemet_stations]{aemet_stations()}}.}

\item{start}{Numeric value with the start year (format: \code{YYYY}).}

\item{end}{Numeric value with the end year (format: \code{YYYY}).}

\item{labels}{Character string with month labels for the x-axis: \code{"en"}
(English), \code{"es"} (Spanish), \code{"fr"} (French), etc.}

\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}

\item{ggplot2}{Logical. If \code{TRUE}, the function uses
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}. If \code{FALSE}, it uses \code{\link[climatol:diagwl]{climatol::diagwl()}}.}

\item{...}{Further arguments passed to
\code{\link[climatol:diagwl]{climatol::diagwl()}} or \code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}, depending on the
value of \CRANpkg{ggplot2}.}
}
\value{
A plot.
}
\description{
Plot a Walter & Lieth climatic diagram from monthly climatology values for
a station. This climatogram is a great way to show a summary of climate
conditions for a place over a specific time period.
}
\note{
The code is based on code from the CRAN package \CRANpkg{climatol}.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
\donttest{
climatogram_period("9434", start = 2015, end = 2020, labels = "en")
}
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Walter, H. K., Harnickell, E., Lieth, F. H. H., & Rehder, H. (1967).
\emph{Klimadiagramm-weltatlas}. Jena: Fischer, 1967.
\item Guijarro J. A. (2023).
\emph{climatol: Climate Tools (Series Homogenization and Derived Products)}. R
package version 4.0.0, \url{https://climatol.eu}.
}
}
\seealso{
Plotting functions:
\code{\link[=climatestripes_station]{climatestripes_station()}},
\code{\link[=climatogram_normal]{climatogram_normal()}},
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}},
\code{\link[=ggstripes]{ggstripes()}},
\code{\link[=ggwindrose]{ggwindrose()}},
\code{\link[=windrose_days]{windrose_days()}},
\code{\link[=windrose_period]{windrose_period()}}

Climatogram functions:
\code{\link{climaemet_9434_climatogram}},
\code{\link[=climatogram_normal]{climatogram_normal()}},
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}}
}
\concept{aemet_plots}
\concept{climatogram}
