% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\encoding{UTF-8}
\name{first_day_of_year}
\alias{first_day_of_year}
\alias{last_day_of_year}
\title{First and last day of year}
\usage{
first_day_of_year(year = NULL)

last_day_of_year(year = NULL)
}
\arguments{
\item{year}{Numeric value as year (format: YYYY).}
}
\value{
Character string as date (format: YYYY-MM-DD).
}
\description{
Get first and last day of year.
}
\examples{
first_day_of_year(2000)
last_day_of_year(2020)
}
\seealso{
Other helpers:
\code{\link[=climaemet_news]{climaemet_news()}},
\code{\link[=dms2decdegrees]{dms2decdegrees()}}
}
\concept{helpers}
