% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\encoding{UTF-8}
\name{dms2decdegrees}
\alias{dms2decdegrees}
\alias{dms2decdegrees_2}
\title{Convert DMS coordinates to decimal degrees}
\usage{
dms2decdegrees(input = NULL)

dms2decdegrees_2(input = NULL)
}
\arguments{
\item{input}{A character string containing DMS coordinates.}
}
\value{
A numeric value.
}
\description{
Converts degrees, minutes and seconds to decimal degrees.
}
\note{
Code for \code{\link[=dms2decdegrees]{dms2decdegrees()}} was adapted from the project at
\url{https://github.com/SevillaR/aemet}.
}
\examples{
dms2decdegrees("055245W")
dms2decdegrees_2("-3º 40' 37\"")
}
\seealso{
Helper functions:
\code{\link[=climaemet_news]{climaemet_news()}},
\code{\link[=first_day_of_year]{first_day_of_year()}}
}
\concept{helpers}
