% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{dms2decdegrees}
\alias{dms2decdegrees}
\title{Converts dms to decimal degrees}
\usage{
dms2decdegrees(input)
}
\arguments{
\item{input}{Character string as DMS coordinates.}
}
\value{
a numeric value.
}
\description{
Converts degrees, minutes and seconds to decimal degrees.
}
\note{
Code modified from project https://github.com/SevillaR/aemet
}
\examples{
\donttest{
dms2decdegrees("055245W")
}

}
