% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dms2decdegrees}
\alias{dms2decdegrees}
\alias{dms2decdegrees_2}
\title{Converts dms to decimal degrees}
\usage{
dms2decdegrees(input = NULL)

dms2decdegrees_2(input = NULL)
}
\arguments{
\item{input}{Character string as DMS coordinates.}
}
\value{
A numeric value.
}
\description{
Converts degrees, minutes and seconds to decimal degrees.
}
\note{
Code for \code{dms2decdegrees()} modified from project
\url{https://github.com/SevillaR/aemet}.
}
\examples{
dms2decdegrees("055245W")
dms2decdegrees_2("-3º 40' 37\"")
}
\seealso{
Other helpers: 
\code{\link{climaemet_news}()},
\code{\link{first_day_of_year}()}
}
\concept{helpers}
