% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dms2decdegrees}
\alias{dms2decdegrees}
\title{Converts dms to decimal degrees}
\usage{
dms2decdegrees(input = NULL)
}
\arguments{
\item{input}{Character string as DMS coordinates.}
}
\value{
a numeric value.
}
\description{
Converts degrees, minutes and seconds to decimal degrees.
}
\note{
Code modified from project \url{https://github.com/SevillaR/aemet}
}
\examples{
dms2decdegrees("055245W")
}
\concept{helpers}
