% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{first_day_of_year}
\alias{first_day_of_year}
\title{First day of year}
\usage{
first_day_of_year(year)
}
\arguments{
\item{year}{Numeric value as year (format: \%Y).}
}
\value{
Character string as date (format: \%Y\%m\%d).
}
\description{
Get first day of year.
}
\examples{
\donttest{
first_day_of_year(2000)
}

}
