% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet-api-query.R
\encoding{UTF-8}
\name{get_data_aemet}
\alias{get_data_aemet}
\alias{get_metadata_aemet}
\title{Query the AEMET API}
\source{
\url{https://opendata.aemet.es/dist/index.html}.
}
\usage{
get_data_aemet(apidest, verbose = FALSE)

get_metadata_aemet(apidest, verbose = FALSE)
}
\arguments{
\item{apidest}{Character string with a destination URL. See
\url{https://opendata.aemet.es/dist/index.html}.}

\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}
}
\value{
A \link[dplyr:tibble]{tibble} (if possible) or the results of the query as
provided by \code{\link[httr2:resp_body_raw]{httr2::resp_body_raw()}} or \code{\link[httr2:resp_body_string]{httr2::resp_body_string()}}.
}
\description{
Retrieves data and metadata from AEMET and converts JSON responses to a
\link[dplyr:tibble]{tibble} when possible.
}
\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
# Run only when AEMET_API_KEY is detected.

url <- "/api/valores/climatologicos/inventarioestaciones/todasestaciones"

get_data_aemet(url)

# Metadata.

get_metadata_aemet(url)

# Get data from any API endpoint.

# Plain text.

plain <- get_data_aemet("/api/prediccion/nacional/hoy")

cat(plain)

# An image.

image <- get_data_aemet("/api/mapasygraficos/analisis")

# Write and read.
tmp <- tempfile(fileext = ".gif")

writeBin(image, tmp)

gganimate::gif_file(tmp)
\dontshow{\}) # examplesIf}
}
\seealso{
See examples of how to use these functions in
\code{vignette("extending-climaemet", package = "climaemet")}.
}
\concept{aemet_low}
