% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_api_query.R
\name{get_data_aemet}
\alias{get_data_aemet}
\title{Client tool for AEMET API}
\source{
\url{https://opendata.aemet.es/dist/index.html}
}
\usage{
get_data_aemet(apidest, apikey = NULL, verbose = FALSE, ...)
}
\arguments{
\item{apidest}{Character string as destination URL. See
\url{https://opendata.aemet.es/dist/index.html}.}

\item{apikey}{Character string as personal API key. It can be set
on the environment variable AEMET_API_KEY, see \code{\link[=aemet_api_key]{aemet_api_key()}}.}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}

\item{...}{Further arguments of the function.}
}
\value{
A tibble. On error, a tibble with a column \code{apidest_error} showing the
errored apidest and \code{error_message} column is returned.
}
\description{
Client tool to get data from AEMET and convert json to tibble.
}
