% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatogram.R
\encoding{UTF-8}
\name{ggclimat_walter_lieth}
\alias{ggclimat_walter_lieth}
\title{Walter-Lieth climate diagram with \CRANpkg{ggplot2}}
\usage{
ggclimat_walter_lieth(
  dat,
  est = "",
  alt = NA,
  per = NA,
  mlab = "es",
  pcol = "#002F70",
  tcol = "#ff0000",
  pfcol = "#9BAEE2",
  sfcol = "#3C6FC4",
  shem = FALSE,
  p3line = FALSE,
  ...
)
}
\arguments{
\item{dat}{Monthly climate data for which the diagram will be plotted.}

\item{est}{Name of the climatological station.}

\item{alt}{Altitude of the climatological station.}

\item{per}{Period used to compute the averages.}

\item{mlab}{Month labels for the x-axis. Use a 2-digit language code
(\code{"en"}, \code{"es"}, etc.). See \code{\link[readr:locale]{readr::locale()}} for details.}

\item{pcol}{Color for precipitation.}

\item{tcol}{Color for temperature.}

\item{pfcol}{Fill color for probable frosts.}

\item{sfcol}{Fill color for sure frosts.}

\item{shem}{Set to \code{TRUE} for southern hemisphere stations.}

\item{p3line}{Set to \code{TRUE} to draw a supplementary precipitation line
relative to three times the temperature (as suggested by Bogdan Rosca).}

\item{...}{Further graphic arguments.}
}
\value{
A \CRANpkg{ggplot2} object. See \code{help("ggplot2")}.
}
\description{
Plots a Walter-Lieth climate diagram for a station. This function is an
updated version of \code{\link[climatol:diagwl]{climatol::diagwl()}} by Jose A. Guijarro.

\if{html}{\figure{lifecycle-experimental.svg}{options: alt="[Experimental]"}}
}
\details{
See the details in \code{\link[climatol:diagwl]{climatol::diagwl()}}.

Climate data must be passed as a 4 x 12 matrix or \link{data.frame} of monthly
data from January to December. Rows must contain mean precipitation, mean
maximum daily temperature, mean minimum daily temperature and absolute
monthly minimum temperature, in that order.

See \link{climaemet_9434_climatogram} for a sample dataset.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{

library(ggplot2)

wl <- ggclimat_walter_lieth(
  climaemet::climaemet_9434_climatogram,
  alt = "249",
  per = "1981-2010",
  est = "Zaragoza Airport"
)

wl

# Since it is a ggplot object, we can modify it.

wl + theme(
  plot.background = element_rect(fill = "grey80"),
  panel.background = element_rect(fill = "grey70"),
  axis.text.y.left = element_text(
    colour = "black",
    face = "italic"
  ),
  axis.text.y.right = element_text(
    colour = "black",
    face = "bold"
  )
)
}
\references{
\itemize{
\item Walter, H. K., Harnickell, E., Lieth, F. H. H., & Rehder, H. (1967).
\emph{Klimadiagramm-weltatlas}. Jena: Fischer, 1967.
\item Guijarro J. A. (2023).
\emph{climatol: Climate Tools (Series Homogenization and Derived Products)}. R
package version 4.0.0, \url{https://climatol.eu}.
}
}
\seealso{
\code{\link[climatol:diagwl]{climatol::diagwl()}}, \code{\link[readr:locale]{readr::locale()}}

Plotting functions:
\code{\link[=climatestripes_station]{climatestripes_station()}},
\code{\link[=climatogram_normal]{climatogram_normal()}},
\code{\link[=climatogram_period]{climatogram_period()}},
\code{\link[=ggstripes]{ggstripes()}},
\code{\link[=ggwindrose]{ggwindrose()}},
\code{\link[=windrose_days]{windrose_days()}},
\code{\link[=windrose_period]{windrose_period()}}

Climatogram functions:
\code{\link{climaemet_9434_climatogram}},
\code{\link[=climatogram_normal]{climatogram_normal()}},
\code{\link[=climatogram_period]{climatogram_period()}}
}
\concept{aemet_plots}
\concept{climatogram}
