% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatogram.R
\name{ggclimat_walter_lieth}
\alias{ggclimat_walter_lieth}
\title{Walter and Lieth climatic diagram on \code{ggplot2}}
\usage{
ggclimat_walter_lieth(
  dat,
  est = "",
  alt = NA,
  per = NA,
  mlab = "es",
  pcol = "#002F70",
  tcol = "#ff0000",
  pfcol = "#9BAEE2",
  sfcol = "#3C6FC4",
  shem = FALSE,
  p3line = FALSE,
  ...
)
}
\arguments{
\item{dat}{Monthly climatic data for which the diagram will be plotted.}

\item{est}{Name of the climatological station}

\item{alt}{Altitude of the climatological station}

\item{per}{Period on which the averages have been computed}

\item{mlab}{Month labels for the X axis. Use 2-digit language code ("en",
"es", etc.). See \code{\link[readr:locale]{readr::locale()}} for info.}

\item{pcol}{Color pen for precipitation.}

\item{tcol}{Color pen for temperature.}

\item{pfcol}{Fill color for probable frosts.}

\item{sfcol}{Fill color for sure frosts.}

\item{shem}{Set to TRUE for southern hemisphere stations.}

\item{p3line}{Set to TRUE to draw a supplementary precipitation line
referenced to three times the temperature (as suggested by Bogdan Rosca).}

\item{...}{Other graphic parameters}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot of a Walter and Lieth climatic diagram of a station. This function is
an updated version of \code{\link[climatol:diagwl]{climatol::diagwl()}}, by Jose A. Guijarro.
}
\details{
See Details on \code{\link[climatol:diagwl]{climatol::diagwl()}}.

Climatic data must be passed as a 4x12 matrix of monthly (January to
December) data, in the following order:
\itemize{
\item Row 1:  Mean precipitation
\item Row 2: Mean maximum daily temperature
\item Row 3: Mean minimum daily temperature
\item Row 4: Absolute monthly minimum temperature
}
}
\references{
Walter, H., and Lieth, H. 1960. \emph{Klimadiagramm-Weltatlas.}
G. Fischer.
}
\seealso{
\code{\link[climatol:diagwl]{climatol::diagwl()}}, \code{\link[readr:locale]{readr::locale()}}
}
\concept{aemet_plots}
