% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripes.R
\encoding{UTF-8}
\name{ggstripes}
\alias{ggstripes}
\title{Plot warming stripes}
\usage{
ggstripes(
  data,
  plot_type = "stripes",
  plot_title = "",
  n_temp = 11,
  col_pal = "RdBu",
  ...
)
}
\arguments{
\item{data}{A \link{data.frame} with date (\code{year}) and temperature (\code{temp})
variables.}

\item{plot_type}{Plot type. Accepted values are \code{"background"},
\code{"stripes"}, \code{"trend"} or \code{"animation"}.}

\item{plot_title}{Character string to be used for the plot title.}

\item{n_temp}{Number of colors in the palette. Defaults to \code{11}.}

\item{col_pal}{Character string indicating the name of the
\code{\link[=hcl.pals]{hcl.pals()}} color palette to be used for plotting.}

\item{...}{Further arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A \CRANpkg{ggplot2} object. See \code{help("ggplot2")}.
}
\description{
Plots "climate stripes" or "warming stripes" with \CRANpkg{ggplot2}.
These graphics represent temperature change at a location over at least 70
years. Each stripe shows the annual average temperature at that station.
}
\note{
Professor Ed Hawkins of the University of Reading developed the "warming
stripes" concept to communicate climate change risks as simply as possible.
For more details, see
\href{https://showyourstripes.info/}{ShowYourStripes}.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\donttest{
library(ggplot2)

data <- climaemet::climaemet_9434_temp

ggstripes(data, plot_title = "Zaragoza Airport") +
  labs(subtitle = "(1950-2020)")

ggstripes(data, plot_title = "Zaragoza Airport", plot_type = "trend") +
  labs(subtitle = "(1950-2020)")
}
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}} for additional
arguments to \code{ggstripes()} and \link{climaemet_9434_wind}.

Warming stripes:
\code{\link[=climatestripes_station]{climatestripes_station()}}
}
\concept{stripes}
