% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{ggstripes}
\alias{ggstripes}
\title{ggstripes}
\usage{
ggstripes(
  data,
  plot_type = c("background", "stripes", "trend", "animation"),
  plot_title = ""
)
}
\arguments{
\item{data}{a data.frame}

\item{plot_type}{plot type}

\item{plot_title}{text}
}
\value{
a ggplot image
}
\description{
Plots diferent climate stripes (whith labels, background, stripes with line trend and animation)
}
\examples{
\dontrun{
ggstripes(data, plot_type = "background")
}
}
