% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemet_tools.R
\name{last_day_of_year}
\alias{last_day_of_year}
\title{Last day of year}
\usage{
last_day_of_year(year, month = 12)
}
\arguments{
\item{year}{a start year (format: \%Y)}

\item{month}{a start year (format: \%m)}
}
\value{
date (format: \%Y\%m\%d)
}
\description{
Get last day of year
}
\examples{
\donttest{
last_day_of_year(2000)
}

}
