% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\encoding{UTF-8}
\name{windrose_period}
\alias{windrose_period}
\title{Wind rose for a range of years}
\usage{
windrose_period(
  station,
  start = 2000,
  end = 2010,
  n_directions = 8,
  n_speeds = 5,
  speed_cuts = NA,
  col_pal = "GnBu",
  calm_wind = 0,
  legend_title = "Wind speed (m/s)",
  verbose = FALSE
)
}
\arguments{
\item{station}{Character string with station identifier code(s). See
\code{\link[=aemet_stations]{aemet_stations()}}.}

\item{start}{Numeric value with the start year (format: \code{YYYY}).}

\item{end}{Numeric value with the end year (format: \code{YYYY}).}

\item{n_directions}{Numeric value with the number of direction bins to plot
(petals on the rose). Valid values are \code{4}, \code{8} or \code{16}.}

\item{n_speeds}{Number of equally spaced wind speed bins to plot when
\code{speed_cuts} is \code{NA}. Defaults to \code{5}.}

\item{speed_cuts}{Numeric vector with the cut points for the wind speed
intervals, or \code{NA} (default).}

\item{col_pal}{Character string indicating the name of the
\code{\link[=hcl.pals]{hcl.pals()}} color palette to be used for plotting.}

\item{calm_wind}{Upper wind speed limit considered calm. Defaults to \code{0}.}

\item{legend_title}{Character string to be used for the legend title.}

\item{verbose}{Logical. If \code{TRUE}, provides information about the flow of
information between the client and server.}
}
\value{
A \CRANpkg{ggplot2} object. See \code{help("ggplot2")}.
}
\description{
Plots a wind rose showing wind speed and direction at a station over a
time period.
}
\section{API key}{
You need to set your API key globally using \code{\link[=aemet_api_key]{aemet_api_key()}}.
Query timeout can be controlled with
\code{options(climaemet_timeout = 60)} (default value). See
\code{\link[httr2:req_timeout]{httr2::req_timeout()}} for details.
}

\examples{
\dontshow{if (aemet_detect_api_key()) withAutoprint(\{ # examplesIf}
\donttest{
# Do not run this example.
if (FALSE) {
  # Downloading data may take a few minutes.
  windrose_period("9434",
    start = 2000, end = 2010,
    speed_cuts = 4
  )
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aemet_daily_period]{aemet_daily_period()}}

Plotting functions:
\code{\link[=climatestripes_station]{climatestripes_station()}},
\code{\link[=climatogram_normal]{climatogram_normal()}},
\code{\link[=climatogram_period]{climatogram_period()}},
\code{\link[=ggclimat_walter_lieth]{ggclimat_walter_lieth()}},
\code{\link[=ggstripes]{ggstripes()}},
\code{\link[=ggwindrose]{ggwindrose()}},
\code{\link[=windrose_days]{windrose_days()}}

Wind functions:
\code{\link{climaemet_9434_wind}},
\code{\link[=ggwindrose]{ggwindrose()}},
\code{\link[=windrose_days]{windrose_days()}}
}
\concept{aemet_plots}
\concept{wind}
