% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose_period}
\alias{windrose_period}
\title{Windrose (speed/direction) diagram of a station over a time period}
\usage{
windrose_period(
  station,
  start = 2000,
  end = 2010,
  n_directions = 8,
  n_speeds = 5,
  speed_cuts = NA,
  col_pal = "GnBu",
  calm_wind = 0,
  legend_title = "Wind Speed (m/s)",
  verbose = FALSE
)
}
\arguments{
\item{station}{Character string with station identifier code(s)
(see \code{\link[=aemet_stations]{aemet_stations()}}) or "all" for all the stations.}

\item{start}{Numeric value as start year (format: \%Y).}

\item{end}{Numeric value as end year (format: \%Y).}

\item{n_directions}{Numeric value as the number of direction bins to plot
(petals on the rose). The number of directions defaults to 8.}

\item{n_speeds}{Numeric value as the number of equally spaced wind speed
bins to plot. This is used if \code{speed_cuts} is "NA" (default 5).}

\item{speed_cuts}{Numeric vector containing the cut points for the wind
speed intervals, or "NA" (default).}

\item{col_pal}{Character string indicating the name of the
\code{\link[=hcl.pals]{hcl.pals()}} colour palette to be used for plotting, see
\strong{Palette selection}.}

\item{calm_wind}{Numeric value as the upper limit for wind speed that is
considered calm (default 0).}

\item{legend_title}{Character string to be used for the legend title.}

\item{verbose}{Logical TRUE/FALSE. Provides information about the flow of
information between the client and server.}
}
\value{
a \code{ggplot2} object
}
\description{
Plot a windrose showing the wind speed and direction for a station over a
time period.
}
\examples{

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  windrose_period("9434",
    start = 2000, end = 2010,
    speed_cuts = 4
  )
}
}
\seealso{
\code{\link[=ggwindrose]{ggwindrose()}}, \code{\link[=windrose_days]{windrose_days()}}, \code{\link[=aemet_daily_period]{aemet_daily_period()}}
}
\concept{aemet_plots}
